% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{adapt_allpic_with_sigma2maxmodel}
\alias{adapt_allpic_with_sigma2maxmodel}
\title{Adapts allpic (object that contains PIC for all candidate C's and all subsamples) with sigma2_max_model.}
\usage{
adapt_allpic_with_sigma2maxmodel(
  all_pic,
  sigma2_max_model,
  UPDATE1 = update1,
  UPDATE2 = update2
)
}
\arguments{
\item{all_pic}{contains PIC for all candidate C's and all subsamples}

\item{sigma2_max_model}{sigma2 of model with maximum number of groups and factors}

\item{UPDATE1}{option to indicate the number of groupfactors is updated during the algorithm; defults to FALSE}

\item{UPDATE2}{option to indicate the number of common factors is updated during the algorithm; defults to FALSE}
}
\description{
The PIC was until now calculated with a sigma2 specific to the configuration (in terms of number of groups and factors).
Because the method requires sigma2 to be equal over all configurations (see proofs of different papers of Ando/Bai) we replace sigma2 by sigma2 of the configuration
with maximum number of groups and factors (this is the last one that is run).
}
