% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{adapt_X_estimating_less_variables}
\alias{adapt_X_estimating_less_variables}
\title{When we run the algorithm with a different number of observable variables then the number we actually have, we need to reformat X.}
\usage{
adapt_X_estimating_less_variables(
  eclipz = FALSE,
  number_of_variables = aantalvars,
  number_vars_estimated = SCHATTEN_MET_AANTALVARS
)
}
\description{
When we run the algorithm with a different number of observable variables then the number we actually have, we need to reformat X.
}
