% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{adapt_X_to_SCHATTEN_MET_AANTALVARS}
\alias{adapt_X_to_SCHATTEN_MET_AANTALVARS}
\title{When we run the algorithm with a different number of observable variables then the number we actually have, we need to reformat X.}
\usage{
adapt_X_to_SCHATTEN_MET_AANTALVARS(eclipz = FALSE)
}
\description{
When we run the algorithm with a different number of observable variables then the number we actually have, we need to reformat X.
}
