% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{add_pic}
\alias{add_pic}
\title{This function fills in all_pic with the calculated PIC for the current configuration.}
\usage{
add_pic(
  df,
  index_configuration,
  use_robust,
  S,
  k,
  kg,
  number_of_variables,
  number_vars_estimated,
  method_estimate_beta,
  pic_e2
)
}
\arguments{
\item{df}{input data frame}

\item{index_configuration}{index of the configuration of groups and factors}

\item{use_robust}{robust or classical estimation}

\item{S}{number of estimated groups}

\item{k}{estimated number of common factors}

\item{kg}{vector with the estimated number of group specific factors for each group}

\item{number_of_variables}{number of observable variables}

\item{number_vars_estimated}{number of variables from which the coefficients are estimated}

\item{method_estimate_beta}{defines how beta is estimated. Default case is an estimated beta for each individual. Default value is "individual." Possible values are "homogeneous", "group" or "individual".}

\item{pic_e2}{NxT matrix with the error terms}
}
\description{
This function fills in all_pic with the calculated PIC for the current configuration.
}
