% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{add_reduced_pic4}
\alias{add_reduced_pic4}
\title{This function adds a reduced form (C and sigma2_max_model are unknown, and assumed equal to one) of the 4th term of the PIC to df.}
\usage{
add_reduced_pic4(df, TT)
}
\arguments{
\item{df}{input data frame}

\item{TT}{length of time series}
}
\description{
More specifically it calculates sum(groupfactors of group j + TN_factor) for PIC2017.
It also adds the reduced form of the alternative PIC's.
}
