% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{bepaal_robpca}
\alias{bepaal_robpca}
\title{Function that uses robust PCA and estimates robust factors and loadings.}
\usage{
bepaal_robpca(object, aantal_eigenvectoren, ALPHA = ALPHA_ROBPCA, KMAX = 20)
}
\arguments{
\item{object}{input}

\item{aantal_eigenvectoren}{number of eigenvectors to extract}

\item{ALPHA}{parameter in robpca -obsolete-}

\item{KMAX}{-explain-}
}
\description{
Contains Robpca (obsolete) and MacroPCA
}
\details{
Notes for MACROPCA:
KMAX: Different values for kmax give different factors, but the  product lambda\emph{factor stays constant. Note that
this number needs to be big enough, otherwise eigen() will be used.
Variation in k does give different results for lambda}factor

NOTE: this function may (not certain) crash (in case of MACROPCA) when ncol(object) >> nrow(object)
Actually it crashes with specific values of dim(object). For example when dim(object) = c(193,27)
}
