% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{bepaal_robuuste_covmatrix}
\alias{bepaal_robuuste_covmatrix}
\title{Function to determine a robust covariance matrix, using MRCD.}
\usage{
bepaal_robuuste_covmatrix(object, type, ALPHA = ALPHA_COVMRCD)
}
\arguments{
\item{object}{input}

\item{type}{subtitle in plot}

\item{ALPHA}{alpha of MRCD}
}
\description{
Function to determine a robust covariance matrix, using MRCD.
}
