% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{beta_real_heterogroups}
\alias{beta_real_heterogroups}
\title{Helpfunction in create_theta_real() for the option theta_real_heterogeen_groups. (This is the deault option.)}
\usage{
beta_real_heterogroups(
  number_of_variables,
  number_of_groups_real,
  EXTRA_THETA_FACTOR = 1
)
}
\arguments{
\item{number_of_variables}{number of variables}

\item{number_of_groups_real}{real numbr of groups}

\item{EXTRA_THETA_FACTOR}{multiplies coefficients in beta; default = 1}
}
\description{
Helpfunction in create_theta_real() for the option theta_real_heterogeen_groups. (This is the deault option.)
}
