% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{calculate_error_term}
\alias{calculate_error_term}
\title{Calculates the error term Y - X*beta_est - LF - LgFg.}
\usage{
calculate_error_term(
  Y,
  X,
  beta_est,
  g,
  factor_group,
  lambda_group,
  comfactor,
  lambda,
  method_estimate_beta,
  vars,
  vars_est,
  S,
  k,
  kg,
  no_common_factorstructure = FALSE,
  no_group_factorstructure = FALSE
)
}
\arguments{
\item{Y}{Y: the panel data of interest}

\item{X}{X}

\item{beta_est}{estimated values of beta}

\item{g}{Vector with group membership for all individuals}

\item{factor_group}{estimated group specific factors}

\item{lambda_group}{loadings of the estimated group specific factors}

\item{comfactor}{estimated common factors}

\item{lambda}{loadings of the estimated common factors}

\item{method_estimate_beta}{defines how beta is estimated. Default case is an estimated beta for each individual. Default value is "individual." Possible values are "homogeneous", "group" or "individual".}

\item{vars}{number of observable variables}

\item{vars_est}{number of variables that will be included in the algorithm and have their coefficient estimated. This is usually equal to vars}

\item{S}{number of estimated groups}

\item{k}{number of common factors to be estimated}

\item{kg}{number of group specific factors to be estimated}

\item{no_common_factorstructure}{if there is a common factorstructure being estimated}

\item{no_group_factorstructure}{if there is a group factorstructure being estimated}
}
\value{
NxT matrix
}
\description{
Calculates the error term Y - X*beta_est - LF - LgFg.
}
