% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{calculate_FL_group_estimated}
\alias{calculate_FL_group_estimated}
\title{Returns the estimated groupfactorstructure.}
\usage{
calculate_FL_group_estimated(
  LAMBDA_GROUP = lambda_group,
  FACTOR_GROUP = factor_group,
  NN = aantal_N,
  TT = aantal_T,
  number_of_groups = aantalgroepen,
  number_of_group_factors = aantalfactoren_groups,
  number_of_common_factors = aantalfactoren_common,
  num_factors_may_vary = aantalfactors_verschillend_per_group
)
}
\arguments{
\item{LAMBDA_GROUP}{loadings of group factors}

\item{FACTOR_GROUP}{group factors}

\item{NN}{number of individuals}

\item{TT}{length of time series}

\item{number_of_groups}{number of groups estimated}

\item{number_of_group_factors}{number of groupfactors estimated}

\item{number_of_common_factors}{number of common factors estimated}

\item{num_factors_may_vary}{whether or not the number of groupfactors is constant over all groups or not}
}
\value{
list with NjxT matrices
}
\description{
Returns the estimated groupfactorstructure.
}
