% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{calculate_FL_group_estimated2}
\alias{calculate_FL_group_estimated2}
\title{Returns the estimated groupfactorstructure.}
\usage{
calculate_FL_group_estimated2(f, l, j)
}
\arguments{
\item{f}{groupfactors}

\item{l}{grouploadings}

\item{j}{Number of groupfactors that are being used in the calculation.}
}
\description{
This is the same function as calculate_FL_group_estimated(), but with adjustable parameters
}
