% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{calculate_FL_group_real}
\alias{calculate_FL_group_real}
\title{Calculate real groupfactorstructure.}
\usage{
calculate_FL_group_real(
  LAMBDA_GROUP_REAL = lambda_group_real,
  FACTOR_GROUP_REAL = factor_group_real
)
}
\value{
list with NjxT matrices
}
\description{
Calculate real groupfactorstructure.
}
