% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{calculate_lambda_group}
\alias{calculate_lambda_group}
\title{calculates factor loadings of groupfactors}
\usage{
calculate_lambda_group(
  use_robust,
  Y,
  X,
  beta_est,
  factor_group,
  g,
  lambda,
  comfactor,
  method_estimate_factors,
  method_estimate_beta,
  S,
  k,
  kg,
  vars,
  vars_est,
  initialise = FALSE,
  UPDATE1 = FALSE,
  UPDATE2 = FALSE
)
}
\arguments{
\item{use_robust}{TRUE or FALSE: defines using the classical or robust algorithm to estimate beta}

\item{Y}{Y: the panel data of interest}

\item{X}{X}

\item{beta_est}{estimated values of beta}

\item{factor_group}{estimated group specific factors}

\item{g}{Vector with estimated group membership for all individuals}

\item{lambda}{loadings of the estimated common factors}

\item{comfactor}{estimated common factors}

\item{method_estimate_factors}{defines method of robust estimaton of the factors: "macro", "pertmm" or "cz"}

\item{method_estimate_beta}{defines how beta is estimated. Default case is an estimated beta for each individual. Default value is "individual." Possible values are "homogeneous", "group" or "individual".}

\item{S}{number of estimated groups}

\item{k}{number of common factors to be estimated}

\item{kg}{number of group specific factors to be estimated}

\item{vars}{number of observable variables}

\item{vars_est}{number of variables that will be included in the algorithm and have their coefficient estimated. This is usually equal to vars}

\item{initialise}{indicator of being in the initialisation phase}

\item{UPDATE1}{option to indicate the number of groupfactors is updated during the algorithm; defaults to FALSE}

\item{UPDATE2}{option to indicate the number of common factors is updated during the algorithm; defaults to FALSE}
}
\description{
returns object which includes group and id of the individuals
}
