% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{calculate_lgfg}
\alias{calculate_lgfg}
\title{Returns list (with as length the number of groups) with lgfg (product of grouploadings a&nd groupfactors).
Each element of the list with the assumption that all individuals are in the same group k.}
\usage{
calculate_lgfg(
  lambda_group,
  factor_group,
  number_of_groups,
  number_of_group_factors,
  number_of_common_factors,
  num_factors_may_vary
)
}
\arguments{
\item{lambda_group}{grouploadings}

\item{factor_group}{groupfactors}

\item{number_of_groups}{number of groups}

\item{number_of_group_factors}{Vector with length number_of_groups. Every element contains the number of groupfactors of that group.}

\item{number_of_common_factors}{number of common factors}

\item{num_factors_may_vary}{TRUE or FALSE, dependingon whether the number of group factors can differ between groups (when estimating this number)}
}
\description{
This function is used to speed up code.
}
\examples{
calculate_lgfg(lambda_group,factor_group,3,c(3,3,3),0,FALSE)
}
