% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{calculate_mse_beta}
\alias{calculate_mse_beta}
\title{Function to calculate MSE of beta_est.}
\usage{
calculate_mse_beta(
  beta_est,
  beta_true,
  without_intercept = FALSE,
  ABintercept = FALSE,
  NN = aantal_N,
  number_of_variables = aantalvars,
  ABDGP1_local = ABDGP1
)
}
\arguments{
\item{beta_est}{estimated values of beta_est}

\item{beta_true}{real values of beta_est}

\item{without_intercept}{boolean to remove the intercept in the calculation}

\item{ABintercept}{boolean to remove the de facto intercept in the calculation (this is only for DGP 1 & 2, due to how they are defined)}

\item{NN}{number of individuals}

\item{number_of_variables}{number of observable variables}

\item{ABDGP1_local}{gives information about which DGP we use; TRUE of FALSE}
}
\description{
It is calculated with the formula of AndoBai2017: supplementary: p19.
Heterogeneous_coefficients_individuals is the default case.
For DGP 1 & 2: When the true number of variables in X is not equal to the standard of 3 it currently returns NA.
}
