% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{calculate_mse_theta}
\alias{calculate_mse_theta}
\title{Function to calculate MSE of beta.}
\usage{
calculate_mse_theta(
  theta,
  theta_real,
  without_intercept = FALSE,
  ABintercept = FALSE
)
}
\arguments{
\item{theta}{estimated values of beta}

\item{theta_real}{real values of beta}

\item{without_intercept}{boolean to remove the intercept in the calculation}

\item{ABintercept}{boolean to remove the de facto intercept in the calculation (this is only for DGP 1 & 2, due to how they were defined)}
}
\description{
(calculated with formula of AndoBai2017: supplementary: p19).
Notes:
Heterogeneous_coefficients_individuals is the default case.
For DGP 1 & 2: When the number of variables in X is not equal to the standard of 3 it returns NA.
}
