% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{calculate_obj_for_g}
\alias{calculate_obj_for_g}
\title{Calculates objective function for individual i and group k in order to estimate group membership.}
\usage{
calculate_obj_for_g(i, k, errors_virtual, rho_parameters, robust, TT)
}
\arguments{
\item{i}{individual}

\item{k}{group}

\item{errors_virtual}{list with errors for each possible group}

\item{rho_parameters}{median and madn of the calculated error term}

\item{robust}{robust or classical estimation}

\item{TT}{length of time series}
}
\description{
Helpfunction in update_g().
Depends on an not yet established group k ( cannot use lgfg_list)
}
