% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{calculate_theta_alpha}
\alias{calculate_theta_alpha}
\title{Update step of algorithm to obtain new theta (and alpha of Bonhomme/Manresa).}
\usage{
calculate_theta_alpha(optimize_kappa = FALSE, eclipz = FALSE)
}
\arguments{
\item{optimize_kappa}{indicates if kappa has to be optimized or not; defaults to FALSE}
}
\value{
list: 1st element contains matrix (N columns: 1 for each element of the panel data) with estimated beta's.
}
\description{
Update step of algorithm to obtain new theta (and alpha of Bonhomme/Manresa).
}
\examples{
#This function needs several initial parameters to be initialized in order to work on itself.
library(RobClustTimeSeries)
library(tidyverse)
library(robustbase)
lambda_group = RobClustTimeSeries::lambda_group_real_dgp3
factor_group = RobClustTimeSeries::factor_group_real_dgp3
lambda = matrix(0,nrow=1,ncol=300) #0 factors -> but needs placeholder
comfactor = matrix(0,nrow=1,ncol=30) #0 factors -> but needs placeholder
X = RobClustTimeSeries::X_dgp3
Y = RobClustTimeSeries::Y_dgp3
g = RobClustTimeSeries::g_real_dgp3
use_robust = TRUE
b = 1 #> 0 group factors
a = 0 #0 common factors
aantalgroepen = 3
aantalfactoren_groups = c(3,3,3)
aantal_N = 300
aantal_T = 30
aantalvars = 3
SCHATTEN_MET_AANTALVARS = 3
homogeneous_coefficients = FALSE
heterogeneous_coefficients_groups = FALSE
heterogeneous_coefficients_individuals = TRUE
ABDGP1 = FALSE
ABintercept = TRUE
theta = calculate_theta_alpha()[[1]]
}
