% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{calculate_TN_factor}
\alias{calculate_TN_factor}
\title{Helpfunction. Calculates part of the 4th term of the PIC.}
\usage{
calculate_TN_factor(TT, Nj)
}
\arguments{
\item{TT}{length of time series}

\item{Nj}{number of time series in group j}
}
\value{
numeric
}
\description{
Helpfunction. Calculates part of the 4th term of the PIC.
}
