% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{calculate_XB_estimated}
\alias{calculate_XB_estimated}
\title{Calculates (the estimated value of) the matrix X*beta_est.}
\usage{
calculate_XB_estimated(
  NN,
  TT,
  number_of_variables,
  number_vars_estimated,
  method_estimate_beta
)
}
\arguments{
\item{NN}{number of individuals}

\item{TT}{length of time series}

\item{number_of_variables}{number of observable variables}

\item{number_vars_estimated}{number of variables that are included in the algorithm and have their coefficient estimated. This is usually equal to number_of_variables.}

\item{method_estimate_beta}{defines how beta is estimated. Default case is an estimated beta for each individual. Default value is "individual." Possible values are "homogeneous", "group" or "individual".}
}
\description{
Calculates (the estimated value of) the matrix X*beta_est.
}
