% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{calculate_XT_estimated}
\alias{calculate_XT_estimated}
\title{Calculates matrix X*beta (the estimated value of this)}
\usage{
calculate_XT_estimated(number_vars_estimated = SCHATTEN_MET_AANTALVARS)
}
\arguments{
\item{number_vars_estimated}{number of observable variables for which a coeffcient will be estimated during the algorithm.
Usual this is equal to the number of observable variables in X.}
}
\description{
Calculates matrix X*beta (the estimated value of this)
}
