% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{calculate_XT_real}
\alias{calculate_XT_real}
\title{Calculates matrix X*beta (the real value of this)}
\usage{
calculate_XT_real()
}
\description{
Calculates matrix X*beta (the real value of this)
}
