% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{check_stopping_rules}
\alias{check_stopping_rules}
\title{Checks the rules for stopping the algorithm, based on its convergence speed.}
\usage{
check_stopping_rules(
  iteration,
  speed,
  all_OF_values,
  speedlimit = 0.01,
  verbose = FALSE
)
}
\arguments{
\item{iteration}{number of iteration}

\item{speed}{convergence speed}

\item{all_OF_values}{vector containing the values of the objective function from previous iterations}

\item{speedlimit}{if the convergence speed falls under this limit the algorithm stops}

\item{verbose}{if TRUE, more information is printed}
}
\description{
Checks the rules for stopping the algorithm, based on its convergence speed.
}
