% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{clustering_with_robust_distances}
\alias{clustering_with_robust_distances}
\title{Function that puts individuals in a separate "class zero", when their distance to all possible groups is bigger then a certain threshold.}
\usage{
clustering_with_robust_distances(g, number_of_groups, Y)
}
\arguments{
\item{g}{Vector with group membership for all individuals}

\item{number_of_groups}{number of groups}

\item{Y}{Y: the panel data of interest}
}
\value{
numeric vector with the new clustering, now including class zero adjustments
}
\description{
It starts with defining a robust location and scatter (based on Ma & Genton (2000): Highly robust estimation of the autocovariance function).
}
