% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{create_beta_real}
\alias{create_beta_real}
\title{Creates beta_real, which contains the real values of beta (= the coefficients of X)}
\usage{
create_beta_real(
  number_of_variables,
  NN = aantal_N,
  number_of_groups_real = aantalgroepen_real,
  eclipz = FALSE,
  extra_beta_factor = 1
)
}
\arguments{
\item{number_of_variables}{number of observable variables}

\item{NN}{number of individuals}

\item{number_of_groups_real}{number of groups}

\item{eclipz}{boolean to indicate using eClipzdatabase; defaults to FALSE}

\item{extra_beta_factor}{multiplies coefficients in beta; default = 1}
}
\value{
matrix with number of rows equal to number of observable variables + 1 (the first row contains the intercept) and number of culumns
equal to the real number of groups.
}
\description{
beta_real_heterogeen_groups is the default case
}
\examples{
#Decide if beta is common, or specific to groups or individuals: Choose 1 of the following 3.
beta_real_homogeen = FALSE
beta_real_heterogeen_groups = TRUE
beta_real_heterogeen_individueel = FALSE
create_beta_real(3,number_of_groups_real = 3)
}
