% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{create_covMat_crosssectional_dependence}
\alias{create_covMat_crosssectional_dependence}
\title{Function used in generating simulated data with non normal errors.}
\usage{
create_covMat_crosssectional_dependence(parameter, NN)
}
\arguments{
\item{parameter}{amount of cross-sectional dependence}

\item{NN}{number of time series}
}
\value{
NxN covariance matrix
}
\description{
Used to include cross-sectional dependence or serial dependence into the simulated panel data.
}
