% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{create_theta_real}
\alias{create_theta_real}
\title{Creates theta_real, which contains the real values of beta (= the coefficients of X)}
\usage{
create_theta_real(
  aantalvars,
  number_of_real_groups = aantalgroepen_real,
  ECLIPZ = FALSE,
  extra_theta_factor = 1
)
}
\arguments{
\item{aantalvars}{number of observable variables}

\item{number_of_real_groups}{number of groups}

\item{extra_theta_factor}{multiplies coefficients in beta; default = 1}

\item{eclipz}{parameter to indicate using real world Eclipzdataset. Defaults to FALSE.}
}
\value{
matrix with number of rows equal to number of observable variables + 1 (the first row contains the intercept) and number of culumns
equal to the real number of groups.
}
\description{
theta_real_heterogeen_groups is the default case
}
\examples{
create_theta_real(3,number_of_real_groups = 3)
}
