% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{define_C_candidates}
\alias{define_C_candidates}
\title{Defines the candidate values for C.}
\usage{
define_C_candidates()
}
\value{
numeric vector
}
\description{
Defines the candidate values for C.
}
\examples{
define_C_candidates()
}
