% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{define_rho_parameters}
\alias{define_rho_parameters}
\title{Determines parameters of rho-function.}
\usage{
define_rho_parameters(object = NULL)
}
\arguments{
\item{object}{input}
}
\value{
list
}
\description{
Robust updating of group membership is based on a rho function (instead of the non-robust quadratic function) on the norm of the errors.
This requires parameters of location and scale.
They are defined here (currently as median and madn).
This function is applied on the estimated errors: Y - XB - FL - FgLg.
This function is used in update_g().
}
