% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{determine_beta}
\alias{determine_beta}
\title{Helpfunction in estimate_beta() for estimating beta_est.}
\usage{
determine_beta(
  string,
  X_special,
  Y_special,
  use_robust,
  initialisation = FALSE,
  indices = NA,
  optimize_kappa = FALSE,
  TT = length_of_time_series,
  NN = number_of_time_series,
  number_of_variables = number_of_variables_fixedvalue,
  nosetting_local = FALSE
)
}
\arguments{
\item{string}{can have values: "homogeneous" (when one beta_est is estimated for all individuals together) or "heterogeneous" (when beta_est is estimated either groupwise or elementwise)}

\item{X_special}{preprocessed X (observable variables)}

\item{Y_special}{preprocessed Y}

\item{use_robust}{robust or classical estimation}

\item{initialisation}{indicator of being in the initialisation phase}

\item{indices}{individuals for which beta_est is being estimated}

\item{optimize_kappa}{indicates if kappa has to be optimized or not (only relevant for the classical algorithm); defaults to FALSE}

\item{TT}{length of time series}

\item{NN}{number of individuals}

\item{number_of_variables}{number of observable variables}

\item{nosetting_local}{option to remove the recommended setting in lmrob(). It is much faster. Defaults to FALSE.}
}
\description{
Helpfunction in estimate_beta() for estimating beta_est.
}
