% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{determine_theta}
\alias{determine_theta}
\title{Helpfunction in estimate_theta() for estimating theta.}
\usage{
determine_theta(
  string,
  X_special,
  Y_special,
  initialisation = FALSE,
  indices = NA,
  optimize_kappa = FALSE,
  TT = aantal_T,
  NN = aantal_N,
  number_of_variables = aantalvars
)
}
\arguments{
\item{string}{can have values: "homogeen" (when one theta is estimated for all individuals together) or "heterogeen" (when theta is estimated either groupwise or elementwise)}

\item{X_special}{preprocessed X (observable variables)}

\item{Y_special}{preprocessed Y}

\item{initialisation}{indicator of being in the initialisation phase}

\item{indices}{individuals for which theta is being estimated}

\item{optimize_kappa}{indicates if kappa has to be optimized or not; defaults to FALSE}

\item{TT}{length of time series}

\item{NN}{number of individuals}

\item{number_of_variables}{number of observable variables}
}
\description{
Helpfunction in estimate_theta() for estimating theta.
}
