% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_df_results_example.R
\docType{data}
\name{df_results_example}
\alias{df_results_example}
\title{An example for df_results. This dataframe contains the estimators for each configuration.}
\format{
Dataframe with 4 rows (one for each configuration) and 11 columns:
\describe{
\item{S}{number of groups}
\item{k_common}{number of common factors}
\item{k1}{number of group specific factors in group 1}
\item{k2}{number of group specific factors in group 2}
\item{k3}{number of group specific factors in group 3}
\item{g}{estimated group membership}
\item{beta_est}{estimated beta}
\item{factor_group}{estimated group specific factors}
\item{lambda_group}{estimated loadings to the group specific factors}
\item{comfactor}{estimated common factors}
\item{lambda_group}{estimated loadings to the common factors}
}
}
\usage{
df_results_example
}
\description{
An example for df_results. This dataframe contains the estimators for each configuration.
}
\keyword{datasets}
