% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{estimate_factor}
\alias{estimate_factor}
\title{Estimates common factor F}
\usage{
estimate_factor(
  use_robust,
  Y,
  X,
  beta_est,
  g,
  lgfg_list,
  method_estimate_factors,
  method_estimate_beta,
  number_of_common_factors,
  number_of_variables,
  number_vars_estimated,
  number_of_group_factors,
  initialise = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{use_robust}{TRUE or FALSE: defines using the classical or robust algorithm to estimate beta}

\item{Y}{Y}

\item{X}{X}

\item{beta_est}{estimated values of beta}

\item{g}{vector with group membership}

\item{lgfg_list}{This is a list (length number of groups) containing FgLg for every group.}

\item{method_estimate_factors}{defines method of robust estimaton of the factors: "macro", "pertmm" or "cz"}

\item{method_estimate_beta}{defines how beta is estimated. Default case is an estimated beta for each individual. Default value is "individual." Possible values are "homogeneous", "group" or "individual".}

\item{number_of_common_factors}{number of common factors to be estimated}

\item{number_of_variables}{number of observable variables}

\item{number_vars_estimated}{number of variables that are included in the algorithm and have their coefficient estimated. This is usually equal to number_of_variables.}

\item{number_of_group_factors}{number of groupfactors to be estimated}

\item{initialise}{boolean}

\item{verbose}{when TRUE, it prints messages}
}
\value{
r x T matrix
}
\description{
The estimator for F, see Anderson (1984), is equal to the first r eigenvectors (multiplied by sqrt(T) due to the restriction F'F/T = I)
associated with first r largest eigenvalues of the matrix WW' (which is TxT)
W is called Z in Ando and Bai (2017)
}
