% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{estimate_theta}
\alias{estimate_theta}
\title{Estimates beta.}
\usage{
estimate_theta(
  optimize_kappa = FALSE,
  eclipz = FALSE,
  NN = aantal_N,
  TT = aantal_T,
  number_of_groups = aantalgroepen,
  number_of_group_factors = aantalfactoren_groups,
  number_of_common_factors = aantalfactoren_common,
  number_of_variables = aantalvars,
  number_vars_estimated = SCHATTEN_MET_AANTALVARS,
  num_factors_may_vary = aantalfactors_verschillend_per_group
)
}
\arguments{
\item{optimize_kappa}{indicates if kappa has to be optimized or not; defaults to FALSE}

\item{eclipz}{parameter to indicate using real world Eclipzdataset. Defaults to FALSE.}

\item{NN}{number of individuals}

\item{TT}{length of time series}

\item{number_of_groups}{number of groups estimated}

\item{number_of_group_factors}{number of groupfactors estimated}

\item{number_of_common_factors}{number of common factors estimated}

\item{number_of_variables}{number of observable variables}

\item{number_vars_estimated}{number of variables that are included in the algorithm and have their coefficient estimated. This is usually equal to number_of_variables.}

\item{num_factors_may_vary}{whether or not the number of groupfactors is constant over all groups or not}
}
\value{
list: 1st element contains matrix (N columns: 1 for each element of the panel data) with estimated theta's.
}
\description{
Update step of algorithm to obtain new estimation for beta. Note that we call it theta because beta() exists in base R.
}
\examples{
#This function needs several initial parameters to be initialized in order to work on itself.
library(RobClustTimeSeries)
library(tidyverse)
library(robustbase)
lambda_group = RobClustTimeSeries::lambda_group_real_dgp3
factor_group = RobClustTimeSeries::factor_group_real_dgp3
lambda = matrix(0,nrow=1,ncol=300) #0 factors -> but needs placeholder
comfactor = matrix(0,nrow=1,ncol=30) #0 factors -> but needs placeholder
X = RobClustTimeSeries::X_dgp3
Y = RobClustTimeSeries::Y_dgp3
g = RobClustTimeSeries::g_real_dgp3
use_robust = TRUE
b = 1 #> 0 group factors
a = 0 #0 common factors
homogeneous_coefficients = FALSE
heterogeneous_coefficients_groups = FALSE
heterogeneous_coefficients_individuals = TRUE
ABDGP1 = FALSE
ABintercept = TRUE
theta = estimate_theta(NN = 300, TT = 30,
number_of_groups = 3, number_of_group_factors = c(3,3,3), number_of_common_factors = 0,
number_of_variables = 3,number_vars_estimated=3,num_factors_may_vary=FALSE)[[1]]
}
