% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{evade_crashes_macropca}
\alias{evade_crashes_macropca}
\title{MacroPCA crashes Rstudio with certain dimensions of the input. Solve this by doubling every row.}
\usage{
evade_crashes_macropca(object)
}
\arguments{
\item{object}{input}
}
\description{
MacroPCA crashes Rstudio with certain dimensions of the input. Solve this by doubling every row.
}
