% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_cleaning.R
\name{evade_floating_point_errors}
\alias{evade_floating_point_errors}
\title{Function to evade floating point errors.}
\usage{
evade_floating_point_errors(A, LIMIT = 1e-13)
}
\arguments{
\item{A}{input}

\item{LIMIT}{limit under which value is set to 0}
}
\description{
Sets values that should be zero but are >0 (e.g. 1e-13) on zero.
}
