% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_factor_group_true_dgp3.R
\docType{data}
\name{factor_group_true_dgp3}
\alias{factor_group_true_dgp3}
\title{factor_group_true_dgp3 contains the values of the true group factors on which Y_dgp3 is based}
\format{
list with length 3: each element has dimension 3 x 30
}
\usage{
factor_group_true_dgp3
}
\description{
factor_group_true_dgp3 contains the values of the true group factors on which Y_dgp3 is based
}
\keyword{datasets}
