% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{final_estimations_filter_kg}
\alias{final_estimations_filter_kg}
\title{Filters dataframe on the requested group specific factors configuration.}
\usage{
final_estimations_filter_kg(df, kg)
}
\arguments{
\item{df}{input dataframe}

\item{kg}{vector with number of group specific factors for each group, on which should be filtered}
}
\description{
Filters dataframe on the requested group specific factors configuration.
}
