% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_g_real_dgp3.R
\docType{data}
\name{g_real_dgp3}
\alias{g_real_dgp3}
\title{g_real_dgp3 contains the real group memberships to the elements of Y_dgp3}
\format{
vector with 300 elements
}
\usage{
g_real_dgp3
}
\description{
g_real_dgp3 contains the real group memberships to the elements of Y_dgp3
}
\keyword{datasets}
