% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_g_true_dgp3.R
\docType{data}
\name{g_true_dgp3}
\alias{g_true_dgp3}
\title{g_true_dgp3 contains the true group memberships of the elements of Y_dgp3}
\format{
vector with 300 elements
}
\usage{
g_true_dgp3
}
\description{
g_true_dgp3 contains the true group memberships of the elements of Y_dgp3
}
\keyword{datasets}
