% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{get_best_configuration}
\alias{get_best_configuration}
\title{This function finds the first stable interval after the first unstable point. It then defines the value for C for the begin, middle and end of this interval.}
\usage{
get_best_configuration(
  list_vc,
  list_rc,
  list_rcj,
  C_candidates,
  S_cand,
  k_cand,
  kg_cand,
  return_short = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{list_vc}{list with resulting expression(VC^2) for each run}

\item{list_rc}{list with resulting rc for each run}

\item{list_rcj}{list with resulting rcj for each run}

\item{C_candidates}{candidates for C}

\item{S_cand}{candidates for S (number of groups)}

\item{k_cand}{candidates for k (number of common factors)}

\item{kg_cand}{candidates for kg (number of group specific factors)}

\item{return_short}{if TRUE, the function returns the dataframe filtered for several specified potential candidates for C}

\item{verbose}{when TRUE, it prints messages}
}
\description{
This function finds the first stable interval after the first unstable point. It then defines the value for C for the begin, middle and end of this interval.
}
