% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{get_final_clustering}
\alias{get_final_clustering}
\title{Function that returns the final clustering, based on the estimated number of groups and common and group specific factors.}
\usage{
get_final_clustering(df, opt_groups, k, kg, limit_est_groups = 20)
}
\arguments{
\item{df}{input dataframe (this will be df_results_full)}

\item{opt_groups}{the optimal number of groups}

\item{k}{the optimal number of common factors}

\item{kg}{vector with the optimal number of group specific factors}

\item{limit_est_groups}{maximum allowed number of groups that can be estimated}
}
\description{
Function that returns the final clustering, based on the estimated number of groups and common and group specific factors.
}
