% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{get_robust_covmatrix}
\alias{get_robust_covmatrix}
\title{Function to determine a robust covariance matrix, using MRCD.}
\usage{
get_robust_covmatrix(object, ALPHA = ALPHA_COVMRCD)
}
\arguments{
\item{object}{input}

\item{ALPHA}{alpha of MRCD}
}
\description{
Function to determine a robust covariance matrix, using MRCD.
}
