% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{grid_add_variables}
\alias{grid_add_variables}
\title{Function which is used to have a df ("grid") with data (individualindex, timeindex, XT and LF) available.}
\usage{
grid_add_variables(
  grid,
  beta,
  lambda,
  comfactor,
  NN = aantal_N,
  TT = aantal_T,
  number_of_variables = aantalvars,
  number_vars_estimated = SCHATTEN_MET_AANTALVARS,
  number_of_groups = aantalgroepen
)
}
\arguments{
\item{grid}{dataframe containing values for X*beta and LF (product of common factor and its loadings)}

\item{beta}{beta}

\item{lambda}{loadings of the common factors}

\item{comfactor}{common factors}

\item{NN}{number of individuals}

\item{TT}{length of time series}

\item{number_of_variables}{number of observable variables}

\item{number_vars_estimated}{number of variables that are included in the algorithm and have their coefficient estimated. This is usually equal to number_of_variables.}

\item{number_of_groups}{number of groups estimated}
}
\description{
It is used in update_g().
}
