% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_cleaning.R
\name{handleNA}
\alias{handleNA}
\title{Function with as input a dataframe. (this will be "Y" or "to_divide") It filters out rows with NA.}
\usage{
handleNA(df)
}
\arguments{
\item{df}{input}
}
\description{
Function with as input a dataframe. (this will be "Y" or "to_divide") It filters out rows with NA.
}
