% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{initialise_df_pic}
\alias{initialise_df_pic}
\title{This function initialises a dataframe which will contain the PIC for each configuration and for each value of C.}
\usage{
initialise_df_pic()
}
\description{
This function initialises a dataframe which will contain the PIC for each configuration and for each value of C.
}
