% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{initialise_rc}
\alias{initialise_rc}
\title{Initialises rc.}
\usage{
initialise_rc(indices_subset, C_candidates)
}
\arguments{
\item{indices_subset}{all indices of the subsets}

\item{C_candidates}{candidates for C (parameter in PIC)}
}
\description{
This function initialises a data frame which will eventually be filled with the optimized number of common factors for each C and for each subset of the original dataset.
}
