% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{initialise_theta}
\alias{initialise_theta}
\title{Initialisation of estimation of beta.}
\usage{
initialise_theta(eclipz = FALSE)
}
\arguments{
\item{eclipz}{parameter to indicate using real world Eclipzdataset. Defaults to FALSE.}
}
\description{
Note: this needs to be called before the definition of grid.
}
\examples{
library(RobClustTimeSeries)
library(tidyverse)
library(robustbase)
lambda_group = RobClustTimeSeries::lambda_group_real_dgp3
factor_group = RobClustTimeSeries::factor_group_real_dgp3
lambda = matrix(0,nrow=1,ncol=300) #0 factors -> but needs placeholder
comfactor = matrix(0,nrow=1,ncol=30) #0 factors -> but needs placeholder
X = RobClustTimeSeries::X_dgp3
Y = RobClustTimeSeries::Y_dgp3
g = RobClustTimeSeries::g_real_dgp3


aantalvars = 3
SCHATTEN_MET_AANTALVARS = 3
aantalgroepen = 3
homogeneous_coefficients = FALSE
heterogeneous_coefficients_groups = FALSE
heterogeneous_coefficients_individuals = TRUE
aantal_N = 300
aantal_T = 30
use_robust = T
ABDGP1 = FALSE
ABintercept = TRUE
theta_init = initialise_theta()
}
