% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{initialise_X}
\alias{initialise_X}
\title{Creates X (the observable variables) to use in simulations.}
\usage{
initialise_X(NN, TT, number_of_variables = aantalvars)
}
\arguments{
\item{NN}{number of individuals}

\item{TT}{length of time series}

\item{number_of_variables}{number of observable variables}
}
\value{
array with dimensions N x T x number of observable variables
}
\description{
X is an array with dimensions N, T and number of observable variables.
The variables are randomly generated with mean 0 and sd 1.
}
\examples{
initialise_X(300,30)
}
