% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{kg_candidates_expand}
\alias{kg_candidates_expand}
\title{Function that returns the set of combinations of groupfactors for which the algorithm needs to run.}
\usage{
kg_candidates_expand(S, kg_min, kg_max, limit_est_groups = 20)
}
\arguments{
\item{S}{number of groups}

\item{kg_min}{minimum value for number of group specific factors}

\item{kg_max}{minimum value for number of group specific factors}

\item{limit_est_groups}{maximum allowed number of groups that can be estimated}
}
\value{
data.frame where each row contains a possible combination of group specific factors for each of the groups
}
\description{
Function that returns the set of combinations of groupfactors for which the algorithm needs to run.
}
