% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_lambda_group_true_dgp3.R
\docType{data}
\name{lambda_group_true_dgp3}
\alias{lambda_group_true_dgp3}
\title{lambda_group_true_dgp3 contains the values of the loadings to the group factors on which Y_dgp3 is based}
\format{
dataframe with 300 rows. The first 3 columns are the loadings to the factors. The 4th column contains group membership. The fifth column contains an id of the individuals.
}
\usage{
lambda_group_true_dgp3
}
\description{
lambda_group_true_dgp3 contains the values of the loadings to the group factors on which Y_dgp3 is based
}
\keyword{datasets}
