% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_parallel.R
\name{make_df_results_parallel}
\alias{make_df_results_parallel}
\title{Makes a dataframe with information on each configuration.}
\usage{
make_df_results_parallel(x, limit_est_groups = 20)
}
\arguments{
\item{x}{output of the parallel version of the algorithm}

\item{limit_est_groups}{maximum allowed number of groups that can be estimated}
}
\value{
data.frame
}
\description{
Makes a dataframe with information on each configuration.
}
