% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{matrixnorm}
\alias{matrixnorm}
\title{Function to calculate the norm of a matrix.}
\usage{
matrixnorm(mat)
}
\arguments{
\item{mat}{input matrix}
}
\value{
numeric
}
\description{
Function to calculate the norm of a matrix.
}
