% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{OF_vectorized3}
\alias{OF_vectorized3}
\title{Calculates objective function: used in local_search + to determine "best_result".}
\usage{
OF_vectorized3(
  NN,
  TT,
  group_memberships,
  grid,
  beta_est = beta_est,
  lc = lambda,
  fc = comfactor,
  lg = lambda_group,
  fg = factor_group,
  number_of_groups,
  number_of_common_factors,
  number_of_group_factors,
  method_estimate_beta,
  num_factors_may_vary
)
}
\arguments{
\item{NN}{number of individuals}

\item{TT}{length of time series}

\item{group_memberships}{Vector containing the group membership for all individuals.}

\item{grid}{dataframe containing XB, FgLg and FL}

\item{beta_est}{estimated values of beta}

\item{lc}{loadings of estimated common factors}

\item{fc}{estimated common factors}

\item{lg}{estimated grouploadings}

\item{fg}{estimated groupfactors}

\item{number_of_groups}{number of groups estimated}

\item{number_of_common_factors}{number of common factors to be estimated}

\item{number_of_group_factors}{number of groupfactors to be estimated}

\item{method_estimate_beta}{defines how beta is estimated. Default case is an estimated beta for each individual. Default value is "individual." Possible values are "homogeneous", "group" or "individual".}

\item{num_factors_may_vary}{whether or not the number of groupfactors is constant over all groups or not}
}
\description{
Calculates objective function: used in local_search + to determine "best_result".
}
