% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{OF_vectorized_helpfunction3}
\alias{OF_vectorized_helpfunction3}
\title{Helpfunction in OF_vectorized3()}
\usage{
OF_vectorized_helpfunction3(
  i,
  t,
  XBETA,
  LF,
  group_memberships,
  lgfg_list,
  Y,
  kg
)
}
\arguments{
\item{i}{index of individual}

\item{t}{index of time}

\item{XBETA}{matrixproduct of X and beta_est}

\item{LF}{matrixproduct of common factors and its loadings}

\item{group_memberships}{Vector with group membership for all individuals}

\item{lgfg_list}{product of groupfactors and their loadings; list with length the number of groups}

\item{Y}{Y: the panel data of interest}

\item{kg}{vector containing the number of group factors to be estimated for all groups}
}
\description{
Helpfunction in OF_vectorized3()
}
