% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{plot_gtable}
\alias{plot_gtable}
\title{Visualises the confusionmatrix of the group membership of the time series.}
\usage{
plot_gtable(g, g_true)
}
\arguments{
\item{g}{estimated groups}

\item{g_true}{true groups}
}
\description{
Visualises the confusionmatrix of the group membership of the time series.
}
