% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03a_IFE_plot_functions.R
\name{plot_typegevallen}
\alias{plot_typegevallen}
\title{Makes a plot of the (weighted) mean and median income for each group}
\usage{
plot_typegevallen(
  band = FALSE,
  groei = TRUE,
  add_grenzen = TRUE,
  Y_norm = TRUE
)
}
\arguments{
\item{band}{adds sd-bands}

\item{groei}{if FALSE then it uses incomelevels instead of incomegrowth}

\item{add_grenzen}{adds grenzen van de pensioenbijdragen (enkel voor groei == FALSE)}

\item{Y_norm:}{normaliseren van de inkomens(groei)}
}
\description{
Weights are based on the relative number of people that retire in a given year ("minpensioenjaar")
}
