% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_IFE_simulations_functions.R
\name{prepare_dfQ}
\alias{prepare_dfQ}
\title{Makes df with summarystats about \% S and \% k correct}
\usage{
prepare_dfQ(df, cutsize, conditional_k_correct = FALSE)
}
\arguments{
\item{conditional_k_correct}{calculates percentage that number of groupfactors is correct under assumption that number of groups is correct}
}
\description{
Makes df with summarystats about \% S and \% k correct
}
