% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{prepare_to_estimate_factor}
\alias{prepare_to_estimate_factor}
\title{Helpfunction in estimate_factor() for the robust approach.}
\usage{
prepare_to_estimate_factor(W, NN = aantal_N, TT = aantal_T)
}
\description{
It has some options used for testing.
}
