% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RobClustTimeSeries.R
\docType{package}
\name{RobClustTimeSeries}
\alias{RobClustTimeSeries}
\title{RobClustTimeSeries}
\description{
This package is about clustering time series in a robust manner.
The method of Ando & Bai (Clustering Huge Number of Financial Time Series: A Panel Data Approach With High-Dimensional Predictors and Factor Structures)
is extended to make it robust against contamination, a common issue with real world data. In this package the core functions for the robust approach are included.
It also contains a simulated dataset (dataset_Y_dgp3).
}
