% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{scaling_X}
\alias{scaling_X}
\title{Scaling of X.}
\usage{
scaling_X(X, firsttime, eclipz = FALSE, number_of_variables = aantalvars)
}
\arguments{
\item{X}{input}

\item{firsttime}{Scaling before generating Y and before adding outliers -> this is always with mean and sd. If this is FALSE, it indicates that}

\item{eclipz}{parameter to indicate using real world Eclipzdataset. Defaults to FALSE.
we are using the function for a second time, after adding the outliers. In the robust case it uses median and MAD, otherwise again mean and sd.}

\item{number_of_variables}{number of observable variables}
}
\description{
Scaling of X.
}
\examples{
X = initialise_X(300,30, number_of_variables = 3)
use_robust = TRUE
scaling_X(X,TRUE, number_of_variables = 3)
}
