% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{schaling_X}
\alias{schaling_X}
\title{Scaling of X.}
\usage{
schaling_X(X, firsttime, eclipz = FALSE)
}
\arguments{
\item{X}{input}

\item{eclipz}{parameter to indicate using real world Eclipzdataset. Defaults to FALSE.}

\item{firsttime:}{Scaling before generating Y and before adding outliers -> this is always with mean and sd. If this is FALSE, it indicates that
we are using the function for a second time, after adding the outliers. In the robust case it uses median and MAD, otherwise again mean and sd.}
}
\description{
Scaling of X.
}
\examples{
X = initialise_X(300,30)
schaling_X(X,TRUE)
}
