% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{show_gtable}
\alias{show_gtable}
\title{This function prints a table of the estimated vs the true group memberships.
If the true group membership is unknown the result is a table with the estimated group membership counts.}
\usage{
show_gtable(g, g_true)
}
\arguments{
\item{g}{vector with estimated group membership for all individuals}

\item{g_true}{vector of length NN with true group memberships}
}
\description{
This function prints a table of the estimated vs the true group memberships.
If the true group membership is unknown the result is a table with the estimated group membership counts.
}
