% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_IFE_simulations_functions.R
\name{simplot1}
\alias{simplot1}
\title{Makes a plot of ...}
\usage{
simplot1(
  aantal_N_filter,
  FGR_filter,
  LGR_filter,
  error_term_special_filter = 0,
  cutsize = 5e-05,
  n_limit = 50,
  conditional_k_correct = FALSE,
  xlimiet = 0.02
)
}
\description{
Makes a plot of ...
}
