% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{solveFG}
\alias{solveFG}
\title{Helpfunction in update_g(), to calculate solve(FG x t(FG)) x FG}
\usage{
solveFG(
  TT = aantal_T,
  number_of_groups = aantalgroepen,
  number_of_group_factors = aantalfactoren_groups
)
}
\description{
Helpfunction in update_g(), to calculate solve(FG x t(FG)) x FG
}
