% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{tabulate_potential_C}
\alias{tabulate_potential_C}
\title{This function shows the configurations for potential C's of the first stable interval (beginpoint, middlepoint and endpoint)}
\usage{
tabulate_potential_C(
  df,
  runs,
  beginpoint,
  middlepoint_log,
  middlepoint,
  endpoint,
  S_cand
)
}
\arguments{
\item{df}{input dataframe}

\item{runs}{number of panel data sets for which the algorithm has run. If larger than one, the median VC2 is used to determine C.}

\item{beginpoint}{first C of the chosen stable interval}

\item{middlepoint_log}{middle C (on a logscale) of the chosen stable interval}

\item{middlepoint}{middle C of the chosen stable interval}

\item{endpoint}{last C of the chosen stable interval}

\item{S_cand}{candidate number for the number of groups}
}
\description{
This function shows the configurations for potential C's of the first stable interval (beginpoint, middlepoint and endpoint)
}
