% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{test_alternative_PIC}
\alias{test_alternative_PIC}
\title{This function tests alternative PIC's.}
\usage{
test_alternative_PIC(C, term2, term3, term4, TT, alternative_pic_2 = FALSE)
}
\arguments{
\item{C}{determines relative contribution of the penalty terms compared to the estimation error term}

\item{term2}{term2 of PIC}

\item{term3}{term3 of PIC}

\item{term4}{term4 of PIC}

\item{TT}{length of time series}

\item{alternative_pic_2}{modifies PIC: for testing purposes}
}
\description{
This function tests alternative PIC's.
}
