% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{theta_real_heterogroups}
\alias{theta_real_heterogroups}
\title{Helpfunction in create_theta_real() for the option theta_real_heterogeen_groups. (This is the deault option)}
\usage{
theta_real_heterogroups(
  aantalvars,
  number_of_real_groups,
  EXTRA_THETA_FACTOR = 1
)
}
\description{
Helpfunction in create_theta_real() for the option theta_real_heterogeen_groups. (This is the deault option)
}
