% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{update_g}
\alias{update_g}
\title{Function that estimates group membership.}
\usage{
update_g()
}
\value{
list: 1st element contains group membership and second element contains the values which are used to determine group membership
}
\description{
Function that estimates group membership.
}
\examples{
#This function needs several initial parameters to be initialized in order to work on itself.
library(RobClustTimeSeries)
library(tidyverse)
library(robustbase)
lambda_group = RobClustTimeSeries::lambda_group_real_dgp3
factor_group = RobClustTimeSeries::factor_group_real_dgp3
lambda = matrix(0,nrow=1,ncol=300) #0 factors -> but needs placeholder
comfactor = matrix(0,nrow=1,ncol=30) #0 factors -> but needs placeholder
X = RobClustTimeSeries::X_dgp3
Y = RobClustTimeSeries::Y_dgp3
g = RobClustTimeSeries::g_real_dgp3
use_robust = TRUE
b = 1 #> 0 group factors
a = 0 #0 common factors
aantalgroepen = 3
aantalfactoren_common = 0
aantalfactoren_groups = c(3,3,3)
aantal_N = 300
aantal_T = 30
grid = expand.grid(1:aantal_N,1:aantal_T)
aantalvars = 3
SCHATTEN_MET_AANTALVARS = 3
homogeneous_coefficients = FALSE
heterogeneous_coefficients_groups = FALSE
heterogeneous_coefficients_individuals = TRUE
eclipz = FALSE
ABDGP1 = FALSE
ABintercept = TRUE
theta = calculate_theta_alpha()[[1]]
grid = grid_add_variables(grid,theta, lambda, comfactor)
g_new = update_g()[[1]]
}
