% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{use_median_values_theta}
\alias{use_median_values_theta}
\title{Replace individually estimated values for theta by the median value within the estimated group}
\usage{
use_median_values_theta(theta, g, number_of_groups)
}
\arguments{
\item{theta}{theta}

\item{g}{vector with group membership}

\item{number_of_groups}{number of groups estimated}
}
\value{
Updated estimation of theta. The format is 1 column for each individual.
}
\description{
Replace individually estimated values for theta by the median value within the estimated group
}
