% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_X_dgp3.R
\docType{data}
\name{X_dgp3}
\alias{X_dgp3}
\title{X_dgp3 contains the values of the observable variables on which Y_dgp3 is based}
\format{
array with 300 x 30 x 3 elements
}
\usage{
X_dgp3
}
\description{
X_dgp3 contains the values of the observable variables on which Y_dgp3 is based
}
\keyword{datasets}
