% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{adapt_pic_with_sigma2maxmodel}
\alias{adapt_pic_with_sigma2maxmodel}
\title{Adapts the object that contains PIC for all candidate C's and all subsamples with sigma2_max_model.}
\usage{
adapt_pic_with_sigma2maxmodel(
  df,
  df_results,
  sigma2_max_model,
  UPDATE1 = FALSE,
  UPDATE2 = FALSE
)
}
\arguments{
\item{df}{contains PIC for all candidate C's and all subsamples}

\item{df_results}{dataframe with results for each estimated configuration}

\item{sigma2_max_model}{sigma2 of model with maximum number of groups and factors}

\item{UPDATE1}{option to indicate the number of groupfactors is updated during the algorithm; defaults to FALSE}

\item{UPDATE2}{option to indicate the number of common factors is updated during the algorithm; defaults to FALSE}
}
\description{
The PIC is calculated with a sigma2 specific to the configuration (= number of groups and factors).
Because the method to estimate the number of groups and factors requires sigma2 to be equal over all configurations
(see proofs of different papers of Ando/Bai) we replace sigma2
by the sigma2 of the configuration with maximum number of groups and factors (this is the last one that was executed).
}
