% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{adapt_X_estimating_less_variables}
\alias{adapt_X_estimating_less_variables}
\title{When we run the algorithm with a different number of observable variables then the number we actually have, we need to reformat X.}
\usage{
adapt_X_estimating_less_variables(
  number_of_variables,
  number_vars_estimated,
  eclipz = FALSE
)
}
\arguments{
\item{number_of_variables}{number of observable variables}

\item{number_vars_estimated}{number of variables from which the coefficients are estimated}

\item{eclipz}{boolean to indicate using eClipzdatabase; defaults to FALSE}
}
\description{
When we run the algorithm with a different number of observable variables then the number we actually have, we need to reformat X.
}
