% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{adapt_X_estimating_less_variables}
\alias{adapt_X_estimating_less_variables}
\title{When running the algorithm with a different number of observable variables then the number that is available, reformat X. (Mainly used for testing)}
\usage{
adapt_X_estimating_less_variables(X, vars_est)
}
\arguments{
\item{X}{dataframe with the observed variables}

\item{vars_est}{number of available observed variables for which a coefficient will be estimated}
}
\value{
Returns a 3D-array. If vars_est is set to 0, it returns NA.
}
\description{
When running the algorithm with a different number of observable variables then the number that is available, reformat X. (Mainly used for testing)
}
