% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{adapt_X_estimating_less_variables}
\alias{adapt_X_estimating_less_variables}
\title{When we run the algorithm with a different number of observable variables then the number we actually have, we need to reformat X.}
\usage{
adapt_X_estimating_less_variables(X, vars, vars_est)
}
\arguments{
\item{X}{dataframe with the observed variables}

\item{vars}{number of observable variables}

\item{vars_est}{number of variables from which the coefficients are estimated}
}
\description{
When we run the algorithm with a different number of observable variables then the number we actually have, we need to reformat X.
}
