% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{add_configuration}
\alias{add_configuration}
\title{This function add the current configuration (number of groups and factors) to df_results}
\usage{
add_configuration(df_results, S, k, kg, limit_est_groups = 20)
}
\arguments{
\item{df_results}{dataframe with results for each estimated configuration}

\item{S}{estimated number of groups in current configuration}

\item{k}{estimated number of common factors in current configuration}

\item{kg}{estimated number of group specific factors in current configuration}

\item{limit_est_groups}{maximum allowed number of groups that can be estimated}
}
\description{
This function add the current configuration (number of groups and factors) to df_results
}
