% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_IFE_robust_lambda.R
\name{all_almost_classical_lambda}
\alias{all_almost_classical_lambda}
\title{calculates vector of which the mean is equal to classical lambda}
\usage{
all_almost_classical_lambda(
  Y_like_object,
  factor_like_object,
  NUMBER_FACTORS,
  group,
  NUMBER_PERSONS
)
}
\value{
matrix where the mean of each row is equal to the classical lambda.
}
\description{
calculates vector of which the mean is equal to classical lambda
}
