% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{beta_true_heterogroups}
\alias{beta_true_heterogroups}
\title{Helpfunction in create_true_beta() for the option beta_true_heterogeen_groups. (This is the default option.)}
\usage{
beta_true_heterogroups(
  number_of_variables,
  number_of_groups_real,
  EXTRA_BETA_FACTOR = 1,
  limit_true_groups_BTH = LIMIT_TRUE_GROUPS
)
}
\arguments{
\item{number_of_variables}{number of observable variables}

\item{number_of_groups_real}{real numbr of groups}

\item{EXTRA_BETA_FACTOR}{option to multiply the coefficients in true beta; default = 1}

\item{limit_true_groups_BTH}{Maximum number of true groups in a simulation-DGP for which the code in this package is implemented. Currently equals 12. For application on realworld data this parameter is not relevant.}
}
\description{
Helpfunction in create_true_beta() for the option beta_true_heterogeen_groups. (This is the default option.)
}
