% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{calculate_error_term}
\alias{calculate_error_term}
\title{Calculates the error term Y - X*BETA - LF - LgFg.}
\usage{
calculate_error_term(
  no_common_factorstructure = FALSE,
  no_group_factorstructure = FALSE
)
}
\value{
NxT matrix
}
\description{
Calculates the error term Y - X*BETA - LF - LgFg.
}
