% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{calculate_error_term}
\alias{calculate_error_term}
\title{Calculates the error term Y - X*beta_est - LF - LgFg.}
\usage{
calculate_error_term(
  Y,
  X,
  beta_est,
  g,
  factor_group,
  lambda_group,
  comfactor,
  lambda,
  method_estimate_beta,
  number_of_variables,
  number_vars_estimated,
  number_of_groups,
  number_of_group_factors,
  number_of_common_factors,
  no_common_factorstructure = FALSE,
  no_group_factorstructure = FALSE
)
}
\arguments{
\item{Y}{Y}

\item{X}{X}

\item{beta_est}{estimated values of beta}

\item{g}{vector with group memberships}

\item{factor_group}{estimated group specific factors}

\item{lambda_group}{loadings of the group specific factors}

\item{comfactor}{common factors}

\item{lambda}{loadings of the common factors}

\item{method_estimate_beta}{defines how beta is estimated. Default case is an estimated beta for each individual. Default value is "individual." Possible values are "homogeneous", "group" or "individual".}

\item{number_of_variables}{number of observable variables}

\item{number_vars_estimated}{number of variables that are included in the algorithm and have their coefficient estimated. This is usually equal to number_of_variables.}

\item{number_of_groups}{number of groups estimated}

\item{number_of_group_factors}{number of groupfactors to be estimated}

\item{number_of_common_factors}{number of common factors to be estimated}

\item{no_common_factorstructure}{if there is a common factorstructure being estimated}

\item{no_group_factorstructure}{if there is a group factorstructure being estimated}
}
\value{
NxT matrix#'
}
\description{
Calculates the error term Y - X*beta_est - LF - LgFg.
}
