% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{calculate_error_term_individuals}
\alias{calculate_error_term_individuals}
\title{Calculates the error term Y - X*BETA - LF - LgFg.}
\usage{
calculate_error_term_individuals()
}
\value{
NxT matrix
}
\description{
Gives same value as calculate_error_term().
}
