% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{calculate_errors_virtual_groups}
\alias{calculate_errors_virtual_groups}
\title{Helpfunction for update_g(). Calculates the errors for every virtual group.}
\usage{
calculate_errors_virtual_groups(
  k,
  LF,
  virtual_grouped_factor_structure,
  NN = aantal_N,
  TT = aantal_T,
  number_of_variables = aantalvars
)
}
\arguments{
\item{k}{group}

\item{LF}{NxT-matrix of the commonfactorstructure}

\item{virtual_grouped_factor_structure}{list with length the number of groups; every element of the list contains NxT-matrix}
}
\value{
NxT matrix with errors (=Y - XB - FgLg - FL)
}
\description{
As we are updating group membership, we use the errorterm as objective function to estimate the group. We assume group membership equals 1,...,NG
(with NG the total number of groups) and calculate the error term.
}
