% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{calculate_FL_group_estimated}
\alias{calculate_FL_group_estimated}
\title{Returns the estimated groupfactorstructure.}
\usage{
calculate_FL_group_estimated(
  LAMBDA_GROUP = lambda_group,
  FACTOR_GROUP = factor_group
)
}
\value{
list with NjxT matrices
}
\description{
Returns the estimated groupfactorstructure.
}
