% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{calculate_FL_group_estimated}
\alias{calculate_FL_group_estimated}
\title{Returns the estimated groupfactorstructure.}
\usage{
calculate_FL_group_estimated(lg, fg, g, NN, TT, S, k, kg, num_factors_may_vary)
}
\arguments{
\item{lg}{loadings of estimated group factors}

\item{fg}{estimated group factors}

\item{g}{Vector with estimated group membership for all individuals}

\item{NN}{number of time series}

\item{TT}{length of time series}

\item{S}{number of estimated groups}

\item{k}{number of common factors to be estimated}

\item{kg}{number of group specific factors to be estimated}

\item{num_factors_may_vary}{whether or not the number of groupfactors is constant over all groups or not}
}
\value{
list with NjxT matrices
}
\description{
Returns the estimated groupfactorstructure.
}
