% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{calculate_FL_group_estimated2}
\alias{calculate_FL_group_estimated2}
\title{Returns the estimated groupfactorstructure.}
\usage{
calculate_FL_group_estimated2(
  f,
  l,
  j,
  number_of_groups = aantalgroepen,
  number_of_common_factors = aantalfactoren_common,
  num_factors_may_vary = aantalfactors_verschillend_per_group,
  NN = aantal_N,
  TT = aantal_T
)
}
\arguments{
\item{f}{groupfactors}

\item{l}{grouploadings}

\item{j}{Number of groupfactors that are being used in the calculation.}

\item{number_of_groups}{number of groups estimated}

\item{number_of_common_factors}{number of common factors estimated}

\item{num_factors_may_vary}{whether or not the number of groupfactors is constant over all groups or not}

\item{NN}{number of individuals}

\item{TT}{length of time series}
}
\description{
This is the same function as calculate_FL_group_estimated(), but with adjustable parameters
}
