% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{calculate_FL_group_real}
\alias{calculate_FL_group_real}
\title{Calculate true groupfactorstructure.}
\usage{
calculate_FL_group_real(
  NN = aantal_N,
  TT = aantal_T,
  LAMBDA_GROUP_REAL = lambda_group_real,
  FACTOR_GROUP_REAL = factor_group_real,
  number_of_groups_real = aantalgroepen_real,
  number_of_group_factors_real = aantalfactoren_groups_real,
  number_of_common_factors_real = aantalfactoren_common_real,
  num_factors_may_vary = aantalfactors_verschillend_per_group,
  eclipz = FALSE,
  ABDGP1_local = ABDGP1,
  using_bramaticroux = exists("DGP_Bramati_Croux")
)
}
\arguments{
\item{NN}{number of individuals}

\item{TT}{length of time series}

\item{LAMBDA_GROUP_REAL}{true group factor loadings}

\item{FACTOR_GROUP_REAL}{true group factors}

\item{number_of_groups_real}{true number of groups}

\item{number_of_group_factors_real}{true number of group factors for each group}

\item{number_of_common_factors_real}{true number of common factors}

\item{num_factors_may_vary}{whether or not the number of groupfactors is constant over all groups or not}

\item{eclipz}{parameter to indicate using real world Eclipzdataset. Defaults to FALSE.}

\item{ABDGP1_local}{gives information about which DGP we use; TRUE of FALSE}

\item{using_bramaticroux}{parameter to indicate that we are using data generated with dgp 5}
}
\value{
list with NjxT matrices
}
\description{
Calculate true groupfactorstructure.
}
