% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{calculate_FL_group_true}
\alias{calculate_FL_group_true}
\title{Calculate true groupfactorstructure.}
\usage{
calculate_FL_group_true(
  NN = aantal_N,
  TT = aantal_T,
  LAMBDA_GROUP_REAL = lambda_group_true,
  FACTOR_GROUP_REAL = factor_group_true,
  true_number_of_groups = aantalgroepen_real,
  true_number_of_group_factors = aantalfactoren_groups_real,
  true_number_of_common_factors = aantalfactoren_common_real,
  num_factors_may_vary = aantalfactors_verschillend_per_group,
  eclipz = FALSE,
  dgp1_AB_local = dgp1_AB,
  using_bramaticroux = exists("DGP_Bramati_Croux")
)
}
\arguments{
\item{NN}{number of individuals}

\item{TT}{length of time series}

\item{LAMBDA_GROUP_REAL}{true group factor loadings}

\item{FACTOR_GROUP_REAL}{true group factors}

\item{true_number_of_groups}{true number of groups}

\item{true_number_of_group_factors}{true number of group factors for each group}

\item{true_number_of_common_factors}{true number of common factors}

\item{num_factors_may_vary}{whether or not the number of groupfactors is constant over all groups or not}

\item{eclipz}{parameter to indicate using real world Eclipzdataset. Defaults to FALSE.}

\item{dgp1_AB_local}{gives information about which DGP we use; TRUE of FALSE}

\item{using_bramaticroux}{parameter to indicate that we are using data generated with dgp 5}
}
\value{
list with NjxT matrices
}
\description{
Calculate true groupfactorstructure.
}
