% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{calculate_FL_group_true}
\alias{calculate_FL_group_true}
\title{Calculate the true groupfactorstructure.}
\usage{
calculate_FL_group_true(
  lgt,
  fgt,
  g_true,
  NN,
  TT,
  S_true,
  k_true,
  kg_true,
  num_factors_may_vary = TRUE,
  dgp1_AB_local = FALSE
)
}
\arguments{
\item{lgt}{true group factor loadings}

\item{fgt}{true group factors}

\item{g_true}{vector of length NN with true group memberships}

\item{NN}{number of time series}

\item{TT}{length of time series}

\item{S_true}{true number of groups}

\item{k_true}{true number of common factors}

\item{kg_true}{true number of group factors for each group}

\item{num_factors_may_vary}{whether or not the number of groupfactors is constant over all groups or not}

\item{dgp1_AB_local}{gives information about which DGP we use; TRUE of FALSE}
}
\value{
list with NjxT matrices
}
\description{
Calculate the true groupfactorstructure.
}
