% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{calculate_lambda}
\alias{calculate_lambda}
\title{calculates factor loadings of common factors}
\usage{
calculate_lambda(
  theta,
  comfactor,
  g,
  lgfg_list,
  initialise = FALSE,
  AANTALFACTOREN_COMMON = aantalfactoren_common
)
}
\arguments{
\item{theta}{beta}

\item{comfactor}{common factors}

\item{g}{vector with group membership}

\item{lgfg_list}{This is a list (length number of groups) containing FgLg for every group.}

\item{initialise}{boolean}

\item{AANTALFACTOREN_COMMON}{number of common factors}
}
\description{
calculates factor loadings of common factors
}
