% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{calculate_lambda}
\alias{calculate_lambda}
\title{calculates factor loadings of common factors}
\usage{
calculate_lambda(
  use_robust,
  Y,
  X,
  beta_est,
  comfactor,
  g,
  lgfg_list,
  method_estimate_factors,
  method_estimate_beta,
  number_of_common_factors,
  number_of_group_factors,
  number_of_variables,
  number_vars_estimated,
  initialise = FALSE
)
}
\arguments{
\item{use_robust}{TRUE or FALSE: defines using the classical or robust algorithm to estimate beta}

\item{Y}{Y}

\item{X}{X}

\item{beta_est}{estimated values of beta}

\item{comfactor}{common factors}

\item{g}{vector with group membership}

\item{lgfg_list}{This is a list (length number of groups) containing FgLg for every group.}

\item{method_estimate_factors}{defines method of robust estimaton of the factors: "macro", "pertmm" or "cz"}

\item{method_estimate_beta}{defines how beta is estimated. Default case is an estimated beta for each individual. Default value is "individual." Possible values are "homogeneous", "group" or "individual".}

\item{number_of_common_factors}{number of common factors to be estimated}

\item{number_of_group_factors}{number of groupfactors to be estimated}

\item{number_of_variables}{number of observable variables}

\item{number_vars_estimated}{number of variables that are included in the algorithm and have their coefficient estimated. This is usually equal to number_of_variables.}

\item{initialise}{boolean}
}
\description{
calculates factor loadings of common factors
}
