% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{calculate_lambda}
\alias{calculate_lambda}
\title{calculates factor loadings of common factors}
\usage{
calculate_lambda(
  beta,
  comfactor,
  g,
  lgfg_list,
  initialise = FALSE,
  NN = aantal_N,
  TT = aantal_T,
  number_of_common_factors = aantalfactoren_common
)
}
\arguments{
\item{beta}{beta}

\item{comfactor}{common factors}

\item{g}{vector with group membership}

\item{lgfg_list}{This is a list (length number of groups) containing FgLg for every group.}

\item{initialise}{boolean}

\item{NN}{number of individuals}

\item{TT}{length of time series}

\item{number_of_common_factors}{number of common factors estimated}
}
\description{
calculates factor loadings of common factors
}
