% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{calculate_lambda}
\alias{calculate_lambda}
\title{calculates factor loadings of common factors}
\usage{
calculate_lambda(
  beta_est,
  comfactor,
  g,
  lgfg_list,
  use_macropca_instead_of_cz,
  efwp = estimate_factors_with_pertMM,
  initialise = FALSE,
  NN = aantal_N,
  TT = aantal_T,
  number_of_common_factors = aantalfactoren_common,
  number_of_variables = aantalvars,
  number_vars_estimated = number_variables_estimated
)
}
\arguments{
\item{beta_est}{beta_est}

\item{comfactor}{common factors}

\item{g}{vector with group membership}

\item{lgfg_list}{This is a list (length number of groups) containing FgLg for every group.}

\item{use_macropca_instead_of_cz}{If TRUE, then factors are estimated robustly with macropca (or pertMM), else with class-zero-method.}

\item{efwp}{indicates that factors are estimated with pertMM}

\item{initialise}{boolean}

\item{NN}{number of individuals}

\item{TT}{length of time series}

\item{number_of_common_factors}{number of common factors to be estimated}

\item{number_of_variables}{number of observable variables}

\item{number_vars_estimated}{number of variables that are included in the algorithm and have their coefficient estimated. This is usually equal to number_of_variables.}
}
\description{
calculates factor loadings of common factors
}
