% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{calculate_lambda}
\alias{calculate_lambda}
\title{calculates factor loadings of common factors}
\usage{
calculate_lambda(
  theta,
  comfactor,
  g,
  lgfg_list,
  use_macropca_instead_of_cz,
  estimate_factors_with_pertMM = estimate_factors_with_pertMM,
  initialise = FALSE,
  NN = aantal_N,
  TT = aantal_T,
  number_of_common_factors = aantalfactoren_common
)
}
\arguments{
\item{theta}{theta}

\item{comfactor}{common factors}

\item{g}{vector with group membership}

\item{lgfg_list}{This is a list (length number of groups) containing FgLg for every group.}

\item{use_macropca_instead_of_cz}{If TRUE, then factors are estimated robustly with macropca (or pertMM), else with class-zero-method.}

\item{estimate_factors_with_pertMM}{indicates that factors are estimated with pertMM}

\item{initialise}{boolean}

\item{NN}{number of individuals}

\item{TT}{length of time series}

\item{number_of_common_factors}{number of common factors estimated}
}
\description{
calculates factor loadings of common factors
}
