% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{calculate_lambda_group}
\alias{calculate_lambda_group}
\title{calculates factor loadings of groupfactors}
\usage{
calculate_lambda_group(
  theta,
  factor_group,
  g,
  lambda,
  comfactor,
  initialise = FALSE,
  UPDATE1 = update1,
  UPDATE2 = update2
)
}
\arguments{
\item{theta}{beta}

\item{factor_group}{group factors}

\item{g}{vector with group membership}

\item{lambda}{loadings}

\item{comfactor}{common factors}

\item{initialise}{boolean}

\item{UPDATE1}{option to indicate the number of groupfactors is updated during the algorithm; defults to FALSE}

\item{UPDATE2}{option to indicate the number of common factors is updated during the algorithm; defults to FALSE}
}
\description{
returns object which includes group and id of the individuals
}
