% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{calculate_lambda_group}
\alias{calculate_lambda_group}
\title{calculates factor loadings of groupfactors}
\usage{
calculate_lambda_group(
  beta_est,
  factor_group,
  g,
  lambda,
  comfactor,
  use_macropca_instead_of_cz,
  initialise = FALSE,
  UPDATE1 = update1,
  UPDATE2 = update2,
  NN = aantal_N,
  TT = aantal_T,
  number_of_groups = aantalgroepen,
  number_of_group_factors = aantalfactoren_groups,
  number_of_common_factors = aantalfactoren_common,
  number_of_variables = aantalvars,
  number_vars_estimated = number_variables_estimated
)
}
\arguments{
\item{beta_est}{beta_est}

\item{factor_group}{group factors}

\item{g}{vector with group membership}

\item{lambda}{loadings}

\item{comfactor}{common factors}

\item{use_macropca_instead_of_cz}{If TRUE, then factors are estimated robustly with macropca (or pertMM), else with class-zero-method.}

\item{initialise}{boolean}

\item{UPDATE1}{option to indicate the number of groupfactors is updated during the algorithm; defaults to FALSE}

\item{UPDATE2}{option to indicate the number of common factors is updated during the algorithm; defaults to FALSE}

\item{NN}{number of individuals}

\item{TT}{length of time series}

\item{number_of_groups}{number of groups estimated}

\item{number_of_group_factors}{number of groupfactors estimated}

\item{number_of_common_factors}{number of common factors estimated}

\item{number_of_variables}{number of observable variables}

\item{number_vars_estimated}{number of variables that are included in the algorithm and have their coefficient estimated. This is usually equal to number_of_variables.}
}
\description{
returns object which includes group and id of the individuals
}
