% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{calculate_lambda_group}
\alias{calculate_lambda_group}
\title{calculates factor loadings of groupfactors}
\usage{
calculate_lambda_group(
  theta,
  factor_group,
  g,
  lambda,
  comfactor,
  use_macropca_instead_of_cz,
  initialise = FALSE,
  UPDATE1 = update1,
  UPDATE2 = update2,
  NN = aantal_N,
  TT = aantal_T,
  number_of_groups = aantalgroepen,
  number_of_group_factors = aantalfactoren_groups
)
}
\arguments{
\item{theta}{theta}

\item{factor_group}{group factors}

\item{g}{vector with group membership}

\item{lambda}{loadings}

\item{comfactor}{common factors}

\item{use_macropca_instead_of_cz}{If TRUE, then factors are estimated with macropca, else with class-zero-method.}

\item{initialise}{boolean}

\item{UPDATE1}{option to indicate the number of groupfactors is updated during the algorithm; defults to FALSE}

\item{UPDATE2}{option to indicate the number of common factors is updated during the algorithm; defults to FALSE}

\item{NN}{number of individuals}

\item{TT}{length of time series}

\item{number_of_groups}{number of groups estimated}

\item{number_of_group_factors}{number of groupfactors estimated}
}
\description{
returns object which includes group and id of the individuals
}
