% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{calculate_lambda_group}
\alias{calculate_lambda_group}
\title{calculates factor loadings of groupfactors}
\usage{
calculate_lambda_group(
  beta,
  factor_group,
  g,
  lambda,
  comfactor,
  initialise = FALSE,
  UPDATE1 = update1,
  UPDATE2 = update2,
  TT = aantal_T,
  number_of_groups = aantalgroepen,
  number_of_group_factors = aantalfactoren_groups
)
}
\arguments{
\item{beta}{beta}

\item{factor_group}{group factors}

\item{g}{vector with group membership}

\item{lambda}{loadings}

\item{comfactor}{common factors}

\item{initialise}{boolean}

\item{UPDATE1}{option to indicate the number of groupfactors is updated during the algorithm; defults to FALSE}

\item{UPDATE2}{option to indicate the number of common factors is updated during the algorithm; defults to FALSE}

\item{TT}{length of time series}

\item{number_of_groups}{number of groups estimated}

\item{number_of_group_factors}{number of groupfactors estimated}
}
\description{
returns object which includes group and id of the individuals
}
