% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{calculate_lgfg}
\alias{calculate_lgfg}
\title{Returns list (with as length the number of groups) with lgfg (product of grouploadings a&nd groupfactors).
Each element of the list with the assumption that all individuals are in the same group k.}
\usage{
calculate_lgfg(
  lambda_group,
  factor_group,
  number_of_groups,
  number_of_group_factors,
  number_of_common_factors,
  num_factors_may_vary,
  NN = aantal_N,
  TT = aantal_T
)
}
\arguments{
\item{lambda_group}{grouploadings}

\item{factor_group}{groupfactors}

\item{number_of_groups}{number of groups estimated}

\item{number_of_group_factors}{number of groupfactors to be estimated}

\item{number_of_common_factors}{number of common factors to be estimated}

\item{num_factors_may_vary}{whether or not the number of groupfactors is constant over all groups or not}

\item{NN}{number of individuals}

\item{TT}{length of time series}
}
\description{
This function is used to speed up code.
}
\examples{
library(tidyverse)
lambda_group = RCTS::lambda_group_real_dgp3
factor_group = RCTS::factor_group_real_dgp3
calculate_lgfg(lambda_group,factor_group,3,c(3,3,3),0,FALSE)
}
