% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{calculate_mse_beta}
\alias{calculate_mse_beta}
\title{Function to calculate MSE of beta_est.}
\usage{
calculate_mse_beta(
  beta_est,
  beta_true,
  without_intercept = FALSE,
  dgp1_spread_group_centers = FALSE,
  NN = number_of_time_series,
  number_of_variables = number_of_variables_fixedvalue,
  dgp1_AB_local = dgp1_AB
)
}
\arguments{
\item{beta_est}{estimated values of beta}

\item{beta_true}{true values of beta}

\item{without_intercept}{boolean to remove the intercept in the calculation}

\item{dgp1_spread_group_centers}{boolean to remove the de facto intercept in the calculation (this is specific for DGP 1, due to how this is defined.)}

\item{NN}{number of individuals}

\item{number_of_variables}{number of observable variables}

\item{dgp1_AB_local}{gives information about which DGP we use; TRUE of FALSE}
}
\description{
It is calculated with the formula of AndoBai2017: supplementary: p19.
Heterogeneous_coefficients_individuals is the default case.
For DGP 1 & 2: When the true number of variables in X is not equal to the standard of 3 it currently returns NA.
}
