% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{calculate_PIC}
\alias{calculate_PIC}
\title{Function to determine PIC (panel information criterium)}
\usage{
calculate_PIC(
  C,
  use_robust,
  number_of_common_factors,
  number_of_group_factors,
  e2,
  sigma2,
  NN = number_of_time_series,
  TT = length_of_time_series,
  number_of_variables = number_of_variables_fixedvalue,
  number_vars_estimated = number_vars_estimated_fixedvalue,
  number_of_groups = number_of_groups_fixedvalue
)
}
\arguments{
\item{C}{determines relative contribution of the penamlty terms}

\item{use_robust}{TRUE or FALSE: defines using the classical, robust algorithm to estimate beta}

\item{number_of_common_factors}{number of common factors to be estimated}

\item{number_of_group_factors}{number of groupfactors to be estimated}

\item{e2}{matrix with error terms}

\item{sigma2}{scalar: sum of squared error terms, scaled by NT}

\item{NN}{number of individuals}

\item{TT}{length of time series}

\item{number_of_variables}{number of observable variables}

\item{number_vars_estimated}{number of variables that are included in the algorithm and have their coefficient estimated. This is usually equal to number_of_variables.}

\item{number_of_groups}{number of groups estimated}
}
\description{
This depends on kappa1 -> kappaN, through p (=number of nonzero elements of beta_est).
The parameter 'sigma2' is the non-robust sigma2. As it is only used in term 2 to 4, it does not actually matter what its value is (needs to be > 0).
Could be set to 1 as well.
}
