% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{calculate_PIC}
\alias{calculate_PIC}
\title{Function to determine PIC (panel information criterium)}
\usage{
calculate_PIC(
  C,
  number_of_common_factors,
  number_of_group_factors,
  e2,
  sigma2,
  number_of_variables = aantalvars,
  number_of_groups = aantalgroepen
)
}
\arguments{
\item{C}{determines relative contribution of the penamlty terms}

\item{number_of_common_factors}{number of common factors}

\item{number_of_group_factors}{vector with number of groupfactors}

\item{e2}{matrix with error terms}

\item{sigma2}{scalar: sum of squared error terms, scaled by NT}
}
\description{
This depends on kappa1 -> kappaN, through p (=number of nonzero elements of theta).
The parameter 'sigma2' is the non-robust sigma2. As it is only used in term 2 to 4, it does not actually matter what its value is (needs to be > 0).
Could be set to 1 as well.
}
