% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{calculate_PIC}
\alias{calculate_PIC}
\title{Function to determine PIC (panel information criterium)}
\usage{
calculate_PIC(C, aantalfactoren_common, aantalfactoren_groups, e2, sigma2)
}
\arguments{
\item{C}{determines relative contribution of the penamlty terms}

\item{aantalfactoren_common}{number of common factors}

\item{aantalfactoren_groups}{vector with number of groupfactors}

\item{e2}{matrix with error terms}

\item{sigma2}{scalar: sum of squared error terms, scaled by NT}
}
\description{
This depends on kappa1 -> kappaN, through p (=number of nonzero elements of theta).
The parameter 'sigma2' is the non-robust sigma2. As it is only used in term 2 to 4, it does not actually matter what its value is (needs to be > 0).
Could be set to 1 as well.
}
