% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{calculate_PIC_term1}
\alias{calculate_PIC_term1}
\title{Function to calculate the first term of PIC (panel information criterium)}
\usage{
calculate_PIC_term1(e2)
}
\arguments{
\item{e2}{NxT matrix with the error terms}
}
\description{
This is used in calculate_PIC()
}
