% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{calculate_PIC_term4}
\alias{calculate_PIC_term4}
\title{Function to calculate term4 of PIC for a group j; includes options for testing alternative PIC's.}
\usage{
calculate_PIC_term4(temp, term4, Nj, NN = aantal_N)
}
\arguments{
\item{temp}{term4 of PIC for group n}

\item{term4}{term 4 of PIC, updated until group n-1; this function updates this value with group n}

\item{Nj}{number of individuals in group j}

\item{NN}{N}
}
\description{
Used in calculate_PIC(), in a loop over all groups.
}
