% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{calculate_PIC_term4}
\alias{calculate_PIC_term4}
\title{Function to calculate term4 of PIC for a group j; includes options for testing alternative PIC's.}
\usage{
calculate_PIC_term4(temp, term4, Nj, NN = aantal_N)
}
\description{
Used in calculate_PIC()
}
