% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{calculate_sigma2}
\alias{calculate_sigma2}
\title{Calculates sum of squared errors, divided by NT}
\usage{
calculate_sigma2(e, NN = nrow(e), TT = ncol(e))
}
\arguments{
\item{e}{matrix with error terms}

\item{NN}{N}

\item{TT}{T}
}
\value{
numeric
}
\description{
Calculates sum of squared errors, divided by NT
}
\examples{
Y <- Y_dgp3
set.seed(1)
e <- matrix(rnorm(nrow(Y) * ncol(Y)), nrow = nrow(Y))
calculate_sigma2(e)
}
