% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{calculate_sigma2}
\alias{calculate_sigma2}
\title{Calculates sum of squared errors, divided by NT}
\usage{
calculate_sigma2(e, NN = aantal_N, TT = aantal_T)
}
\arguments{
\item{e}{matrix with error terms}

\item{NN}{N}

\item{TT}{T}
}
\description{
Calculates sum of squared errors, divided by NT
}
