% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{calculate_sigma2maxmodel}
\alias{calculate_sigma2maxmodel}
\title{Calculates sigma2maxmodel}
\usage{
calculate_sigma2maxmodel(
  optimize_kappa = FALSE,
  UPDATE1 = update1,
  UPDATE2 = update2,
  number_of_groups = aantalgroepen,
  number_of_groups_candidates = aantalgroepen_candidates,
  number_of_group_factors = aantalfactoren_groups,
  number_of_common_factors = aantalfactoren_common
)
}
\arguments{
\item{optimize_kappa}{indicates if kappa has to be optimized or not; defaults to FALSE}

\item{UPDATE1}{option to indicate the number of groupfactors is updated during the algorithm; defults to FALSE}

\item{UPDATE2}{option to indicate the number of common factors is updated during the algorithm; defults to FALSE}

\item{number_of_groups}{number of groups estimated}

\item{number_of_groups_candidates}{vector with candidate values for the number of groups}

\item{number_of_group_factors}{number of groupfactors estimated}

\item{number_of_common_factors}{number of common factors}
}
\description{
Sigma2 is the sum of the squared errors, divided by NT. We need the sigma2 of the maxmodel to use (in term 2,3,4 of the PIC) instead of the configuration-dependent sigma2. (based on text in AndoBai 2016).
sigma2_max_model could actually be set to 1 as well, as it can be absorbed in parameter C of the PIC.
}
