% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{calculate_virtual_factor_and_lambda_group}
\alias{calculate_virtual_factor_and_lambda_group}
\title{Helpfunction used in update_g()}
\usage{
calculate_virtual_factor_and_lambda_group(
  group,
  solve_FG_FG_times_FG,
  NN,
  number_of_common_factors = aantalfactoren_common
)
}
\arguments{
\item{group}{number of group}

\item{solve_FG_FG_times_FG}{This is the same as groupfactor / T. It is only used in the Classical approach}

\item{NN}{number of individuals}

\item{number_of_common_factors}{number of common factors estimated}
}
\value{
NxT matrix containing the product of virtual groupfactors and virtual loadings
}
\description{
We calculate FgLg (the groupfactorstructure) for all  possible groups where individual i can be placed. For each group we have before estimated
the groupfactors (Fg). Now we need to the grouploadings for each group as well. In the classical case these are calculated by Fg*Y/T. In the robust case
these are robust.
}
