% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{calculate_virtual_factor_and_lambda_group}
\alias{calculate_virtual_factor_and_lambda_group}
\title{Helpfunction used in update_g()}
\usage{
calculate_virtual_factor_and_lambda_group(
  group,
  solve_FG_FG_times_FG,
  NN_local,
  TT_local,
  use_macropca_instead_of_cz_local,
  number_of_variables_local = aantalvars,
  number_vars_estimated_local = number_variables_estimated,
  number_of_group_factors_local = aantalfactoren_groups,
  number_of_common_factors_local = aantalfactoren_common,
  eclipz_local = eclipz,
  verbose = FALSE
)
}
\arguments{
\item{group}{number of groups}

\item{solve_FG_FG_times_FG}{This is the same as groupfactor / T. It is only used in the Classical approach}

\item{NN_local}{N}

\item{TT_local}{T}

\item{use_macropca_instead_of_cz_local}{If TRUE, then factors are estimated robustly with macropca (or pertMM), else with class-zero-method.}

\item{number_of_variables_local}{number of observable variables}

\item{number_vars_estimated_local}{number of variables that are included in the algorithm and have their coefficient estimated. This is usually equal to number_of_variables.}

\item{number_of_group_factors_local}{number of group factors to be estimated}

\item{number_of_common_factors_local}{number of common factors to be estimated}

\item{eclipz_local}{Parameter to indicate using real world Eclipzdataset. Defaults to FALSE.}

\item{verbose}{when TRUE, it prints messages}
}
\value{
NxT matrix containing the product of virtual groupfactors and virtual loadings
}
\description{
We calculate FgLg (the groupfactorstructure) for all possible groups where individual i can be placed. For each group we have before estimated
the groupfactors (Fg). Now we need the grouploadings for each group as well. In the classical case these are calculated by Fg*Y/T. In the robust case
these are robust.
}
