% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{calculate_W}
\alias{calculate_W}
\title{Calculates W = Y - X*BETA}
\usage{
calculate_W(theta, g)
}
\arguments{
\item{theta}{beta}

\item{g}{vector with group membership}
}
\value{
NxT matrix
}
\description{
The equivalent matrix is called z in Ando/Bai 2017
}
