% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{calculate_W}
\alias{calculate_W}
\title{Calculates W = Y - X*BETA}
\usage{
calculate_W(
  beta,
  g,
  NN = aantal_N,
  TT = aantal_T,
  number_of_variables = aantalvars,
  number_vars_estimated = SCHATTEN_MET_AANTALVARS,
  eclipz = FALSE
)
}
\arguments{
\item{beta}{beta}

\item{g}{vector with group membership}

\item{NN}{number of individuals}

\item{TT}{length of time series}

\item{number_of_variables}{number of observable variables}

\item{number_vars_estimated}{number of variables that are included in the algorithm and have their coefficient estimated. This is usually equal to number_of_variables.}

\item{eclipz}{parameter to indicate using real world Eclipzdataset. Defaults to FALSE.}
}
\value{
NxT matrix
}
\description{
The equivalent matrix is called z in Ando/Bai 2017
}
