% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{calculate_XB_true}
\alias{calculate_XB_true}
\title{Calculates the product of X*beta_true .}
\usage{
calculate_XB_true(NN, TT, number_of_variables, g = g, method_estimate_beta)
}
\arguments{
\item{NN}{number of individuals}

\item{TT}{length of time series}

\item{number_of_variables}{number of observable variables}

\item{g}{estimated group membership}

\item{method_estimate_beta}{defines how beta is estimated. Default case is an estimated beta for each individual. Default value is "individual." Possible values are "homogeneous", "group" or "individual".}
}
\description{
Calculates the product of X*beta_true .
}
