% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{calculate_XB_true}
\alias{calculate_XB_true}
\title{Calculates the product of X*beta_true .}
\usage{
calculate_XB_true(X, beta_true, g, g_true, method_estimate_beta)
}
\arguments{
\item{X}{X: NxTxp array containing the observable variables}

\item{beta_true}{true coefficients of the observable variables}

\item{g}{Vector with estimated group membership for all individuals}

\item{g_true}{true group membership}

\item{method_estimate_beta}{defines how beta is estimated. Default case is an estimated beta for each individual. Default value is "individual." Possible values are "homogeneous", "group" or "individual".}
}
\value{
Returns a NxT matrix (if method_estimate_beta == "individual"), and otherwise NA.
}
\description{
Calculates the product of X*beta_true .
}
