% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{calculate_Z_common}
\alias{calculate_Z_common}
\title{Calculates Z = Y - X*BETA - LgFg, to use in estimate of common factorstructure}
\usage{
calculate_Z_common(
  theta,
  g,
  lgfg_list,
  initialise = FALSE,
  number_of_variables = aantalvars
)
}
\arguments{
\item{theta}{beta}

\item{g}{vector with group membership}

\item{lgfg_list}{This is a list (length number of groups) containing FgLg for every group.}

\item{initialise}{boolean}
}
\description{
Calculates Z = Y - X*BETA - LgFg, to use in estimate of common factorstructure
}
