% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{calculate_Z_group}
\alias{calculate_Z_group}
\title{Calculates Z = Y - X*beta_est - LF. It is used to estimate the groupfactorstructure.}
\usage{
calculate_Z_group(
  beta_est,
  g,
  lambda,
  comfactor,
  group,
  initialise,
  TT = length_of_time_series,
  number_of_variables = number_of_variables_fixedvalue,
  number_vars_estimated = number_vars_estimated_fixedvalue,
  number_of_common_factors = number_of_common_factors_fixedvalue
)
}
\arguments{
\item{beta_est}{estimated values of beta}

\item{g}{vector with group membership}

\item{lambda}{common factor loadings}

\item{comfactor}{common factors}

\item{group}{the number of the group}

\item{initialise}{boolean}

\item{TT}{length of time series}

\item{number_of_variables}{number of observable variables}

\item{number_vars_estimated}{number of variables that are included in the algorithm and have their coefficient estimated. This is usually equal to number_of_variables.}

\item{number_of_common_factors}{number of common factors to be estimated}
}
\description{
Calculates Z = Y - X*beta_est - LF. It is used to estimate the groupfactorstructure.
}
