% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{calculate_Z_group}
\alias{calculate_Z_group}
\title{Calculates Z = Y - X*THETA - LF), to use in estimate of groupfactorstructure.}
\usage{
calculate_Z_group(
  theta,
  g,
  lambda,
  comfactor,
  group,
  initialise,
  TT = aantal_T,
  number_of_variables = aantalvars,
  number_vars_estimated = SCHATTEN_MET_AANTALVARS,
  eclipz = FALSE,
  number_of_common_factors = aantalfactoren_common
)
}
\arguments{
\item{theta}{theta}

\item{g}{vector with group membership}

\item{lambda}{common factor loadings}

\item{comfactor}{common factors}

\item{group}{the number of the group}

\item{initialise}{boolean}

\item{TT}{length of time series}

\item{number_of_variables}{number of observable variables}

\item{number_vars_estimated}{number of variables that are included in the algorithm and have their coefficient estimated. This is usually equal to number_of_variables.}

\item{eclipz}{parameter to indicate using real world Eclipzdataset. Defaults to FALSE.}

\item{number_of_common_factors}{number of common factors estimated}
}
\description{
Calculates Z = Y - X*THETA - LF), to use in estimate of groupfactorstructure.
}
