% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{calculate_Z_group}
\alias{calculate_Z_group}
\title{Calculates Z = Y - X*BETA - LF), to use in estimate of groupfactorstructure.}
\usage{
calculate_Z_group(
  theta,
  g,
  lambda,
  comfactor,
  group,
  initialise,
  TT = aantal_T,
  number_of_variables = aantalvars
)
}
\arguments{
\item{theta}{beta}

\item{g}{vector with group membership}

\item{lambda}{common factor loadings}

\item{comfactor}{common factors}

\item{group}{the number of the group}

\item{initialise}{boolean}

\item{TT}{length of time series}

\item{number_of_variables}{number of observable variables}
}
\description{
Calculates Z = Y - X*BETA - LF), to use in estimate of groupfactorstructure.
}
