% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{check_stopping_rules}
\alias{check_stopping_rules}
\title{This function checks the rules for stopping the algorithm, based on its convergence speed.}
\usage{
check_stopping_rules(speed, speedlimit = 0.01)
}
\arguments{
\item{speed}{convergence speed}

\item{speedlimit}{if the convergence speed falls under this limit the algorithm stops}
}
\description{
This function checks the rules for stopping the algorithm, based on its convergence speed.
}
