% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{create_data_dgp2}
\alias{create_data_dgp2}
\title{This function creates an instance of DGP 2, as defined in \insertCite{BoudtHeyndels2021;textual}{RCTS}.}
\usage{
create_data_dgp2(
  N,
  TT,
  true_number_of_groups = 3,
  vars = 3,
  true_number_of_common_factors = 0,
  true_number_of_group_factors = c(3, 3, 3)
)
}
\arguments{
\item{N}{number of time series}

\item{TT}{length of time series}

\item{true_number_of_groups}{true number of groups}

\item{vars}{number of available observed variables}

\item{true_number_of_common_factors}{true number of common_factors}

\item{true_number_of_group_factors}{vector with the true number of group factors for each group}
}
\description{
The output is a dataframe with N (amount of time series) rows and T (length of time series) columns.
}
