% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{create_theta_real}
\alias{create_theta_real}
\title{Creates theta_real, which contains the real values of theta (= the coefficients of X)}
\usage{
create_theta_real(
  number_of_variables,
  NN = aantal_N,
  number_of_groups_real = aantalgroepen_real,
  eclipz = FALSE,
  extra_theta_factor = 1
)
}
\arguments{
\item{number_of_variables}{number of observable variables}

\item{NN}{number of individuals}

\item{number_of_groups_real}{number of groups}

\item{eclipz}{boolean to indicate using eClipzdatabase; defaults to FALSE}

\item{extra_theta_factor}{multiplies coefficients in theta; default = 1}
}
\value{
matrix with number of rows equal to number of observable variables + 1 (the first row contains the intercept) and number of culumns
equal to the real number of groups.
}
\description{
theta_real_heterogeen_groups is the default case
}
\examples{
#Decide if theta is common, or specific to groups or individuals: Choose 1 of the following 3.
theta_real_homogeen = FALSE
theta_real_heterogeen_groups = TRUE
theta_real_heterogeen_individueel = FALSE
create_theta_real(3,number_of_groups_real = 3)
}
