% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{create_true_beta}
\alias{create_true_beta}
\title{Creates beta_true, which contains the real values of beta_est (= the coefficients of X)}
\usage{
create_true_beta(
  number_of_variables,
  NN = aantal_N,
  number_of_groups_real = aantalgroepen_real,
  eclipz = FALSE,
  extra_beta_factor = 1,
  beta_true_homogeneous = beta_true_homogeen,
  beta_true_heterogeneous_groups = beta_true_heterogeen_groups,
  beta_true_heterogeneous_individuals = beta_true_heterogeen_individueel,
  limit_true_groups = LIMIT_TRUE_GROUPS
)
}
\arguments{
\item{number_of_variables}{number of observable variables}

\item{NN}{number of individuals}

\item{number_of_groups_real}{number of groups}

\item{eclipz}{boolean to indicate using eClipzdatabase; defaults to FALSE}

\item{extra_beta_factor}{multiplies coefficients in beta_est; default = 1}

\item{beta_true_homogeneous}{whether true beta is equal for all individuals}

\item{beta_true_heterogeneous_groups}{whether true beta is equal within groups, and different between groups}

\item{beta_true_heterogeneous_individuals}{whether true beta is different for all individuals}
}
\value{
matrix with number of rows equal to number of observable variables + 1 (the first row contains the intercept) and number of culumns
equal to the real number of groups.
}
\description{
beta_true_heterogeen_groups is the default case
}
\examples{
library(tidyverse)
#Decide if beta_est is common, or specific to groups or individuals: Choose 1 of the following 3.
create_true_beta(3, NN = 300, number_of_groups_real = 3,
  beta_true_homogeneous = FALSE, beta_true_heterogeneous_groups = TRUE,
  beta_true_heterogeneous_individuals = FALSE, limit_true_groups = 12)
}
