% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{define_C_candidates}
\alias{define_C_candidates}
\title{This function defines the candidate values for C.}
\usage{
define_C_candidates()
}
\description{
This function defines the candidate values for C.
}
