% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{define_number_subsets}
\alias{define_number_subsets}
\title{This function returns a vector with the indices of the subsets. Must start with zero.}
\usage{
define_number_subsets(n)
}
\arguments{
\item{n}{number of subsets}
}
\description{
This function returns a vector with the indices of the subsets. Must start with zero.
}
