% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{define_rho_parameters}
\alias{define_rho_parameters}
\title{Determines parameters of rho-function.}
\usage{
define_rho_parameters(object = NULL)
}
\description{
In robust updating of group membership we use a rho function (instead of the non-robust quadratic function) on the norm of the errors.
For this we need parameters of location and scale.
They are defined here (currently as median and madn).
This function is applied on the estimated errors: Y - XB - FL - FgLg.
This function is used in update_g().
}
\examples{
lambda = ... #estimate common factor loadings
comfactor = ... #estimate common factor
LF = (t(lambda) \%*\% comfactor)
aantalgroepen = 3
solve_FG_FG_times_FG = solveFG()
virtual_grouped_factor_structure = lapply(1:aantalgroepen, function(y) calculate_virtual_factor_and_lambda_group(y, solve_FG_FG_times_FG))
ERRORS_VIRTUAL = lapply(1:aantalgroepen, function(x) calculate_errors_virtual_groups(x,LF,virtual_grouped_factor_structure))
define_rho_parameters(ERRORS_VIRTUAL[[x]]))
}
