% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{determine_beta}
\alias{determine_beta}
\title{Helpfunction in estimate_beta() for estimating theta.}
\usage{
determine_beta(
  string,
  X_special,
  Y_special,
  correct,
  initialisatie = FALSE,
  indices = NA,
  optimize_kappa = FALSE,
  number_of_variables = aantalvars
)
}
\arguments{
\item{string}{can have values: "homogeen" or "heterogeen"}

\item{X_special}{preprocessed X (observable variables)}

\item{Y_special}{preprocessed Y}

\item{correct}{Defaults to TRUE. If TRUE, then My is not being used. My is an scaling thing that came out of Ando/Bai.}

\item{initialisatie}{boolean}

\item{optimize_kappa}{indicates if kappa has to be optimized or not; defaults to FALSE}

\item{indices:}{individuals for which beta is being estimated}
}
\description{
Helpfunction in estimate_beta() for estimating theta.
}
