% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{determine_beta}
\alias{determine_beta}
\title{Helpfunction in estimate_beta() for estimating beta_est.}
\usage{
determine_beta(
  string,
  X_special,
  Y_special,
  use_robust,
  NN,
  TT,
  number_of_groups,
  method_estimate_beta,
  initialisation = FALSE,
  indices = NA,
  number_of_variables,
  nosetting_local = FALSE,
  kappa_candidates = c(0.1),
  special_case_dgp1 = FALSE
)
}
\arguments{
\item{string}{can have values: "homogeneous" (when one beta_est is estimated for all individuals together) or "heterogeneous" (when beta_est is estimated either groupwise or elementwise)}

\item{X_special}{preprocessed X (observable variables)}

\item{Y_special}{preprocessed Y}

\item{use_robust}{robust or classical estimation}

\item{NN}{number of individuals}

\item{TT}{length of time series}

\item{number_of_groups}{number of groups}

\item{method_estimate_beta}{defines how beta is estimated. Default case is an estimated beta for each individual. Default value is "individual." Possible values are "homogeneous", "group" or "individual".}

\item{initialisation}{indicator of being in the initialisation phase}

\item{indices}{individuals for which beta_est is being estimated}

\item{number_of_variables}{number of observable variables}

\item{nosetting_local}{option to remove the recommended setting in lmrob(). It is much faster. Defaults to FALSE.}

\item{kappa_candidates}{defines the size of the SCAD-penalty used in the classical algorithm}

\item{special_case_dgp1}{special case for data generated according to dgp 1: it changes the 1st variable in X to 1 (-> intercept). Consequently the estimation of beta needs to be restructured slightly.}
}
\description{
Helpfunction in estimate_beta() for estimating beta_est.
}
