% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{do_we_estimate_group_factors}
\alias{do_we_estimate_group_factors}
\title{Helpfunction to shorten code: use a and b}
\usage{
do_we_estimate_group_factors(number_of_group_factors)
}
\arguments{
\item{number_of_group_factors}{number of group factors to be estimated}
}
\description{
Helpfunction to shorten code: use a and b
}
