% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{estimate_beta}
\alias{estimate_beta}
\title{Estimates beta.}
\usage{
estimate_beta(
  use_robust,
  Y,
  X,
  lambda_group,
  factor_group,
  lambda,
  comfactor,
  method_estimate_beta = "individual",
  NN = nrow(Y),
  TT = ncol(Y),
  number_of_groups,
  number_of_group_factors,
  number_of_common_factors,
  number_of_variables,
  number_vars_estimated,
  num_factors_may_vary,
  optimize_kappa = FALSE,
  nosetting = FALSE,
  special_case_dgp1 = FALSE
)
}
\arguments{
\item{use_robust}{TRUE or FALSE: defines using the classical or robust algorithm to estimate beta}

\item{Y}{Y}

\item{X}{X}

\item{lambda_group}{loadings of the group specific factors}

\item{factor_group}{estimated group specific factors}

\item{lambda}{loadings of the common factors}

\item{comfactor}{estimated common factors}

\item{method_estimate_beta}{defines how beta is estimated. Default case is an estimated beta for each individual. Default value is "individual." Possible values are "homogeneous", "group" or "individual".}

\item{NN}{number of individuals}

\item{TT}{length of time series}

\item{number_of_groups}{number of groups estimated}

\item{number_of_group_factors}{number of groupfactors to be estimated}

\item{number_of_common_factors}{number of common factors to be estimated}

\item{number_of_variables}{number of observable variables}

\item{number_vars_estimated}{number of variables that are included in the algorithm and have their coefficient estimated. This is usually equal to number_of_variables.}

\item{num_factors_may_vary}{whether or not the number of groupfactors is constant over all groups or not}

\item{optimize_kappa}{indicates if kappa has to be optimized or not (only relevant for the classical algorithm); defaults to FALSE}

\item{nosetting}{option to remove the recommended setting in lmrob(). It is much faster. Defaults to FALSE.}

\item{special_case_dgp1}{special case for data generated according to dgp 1: it changes the 1st variable in X to 1 (-> intercept). Consequently the estimation of beta needs to be restructured slightly.}
}
\value{
list: 1st element contains matrix (N columns: 1 for each element of the panel data) with estimated beta_est's.
}
\description{
Update step of algorithm to obtain new estimation for beta. Note that we call it beta_est because beta() exists in base R.
}
\examples{
#This function needs several initial parameters to be initialized in order to work on itself.
library(RCTS)
library(tidyverse)
library(robustbase)

X = RCTS::X_dgp3
Y = RCTS::Y_dgp3
#Set estimations for group factors and its loadings, and group membership to the true value
lambda_group = RCTS::lambda_group_true_dgp3
factor_group = RCTS::factor_group_true_dgp3
g = RCTS::g_true_dgp3
#There are no common factors to be estimated  -> but needs placeholder
lambda = matrix(0, nrow = 1, ncol = 300)
comfactor = matrix(0, nrow = 1, ncol=30)
#Choose how coefficients of the observable variables are estimated
method_estimate_beta = "individual"
method_estimate_factors = "macro"
beta_est = estimate_beta(use_robust = TRUE, Y, X, lambda_group, factor_group,
  lambda, comfactor,
  number_of_groups = 3, number_of_group_factors = c(3, 3, 3), number_of_common_factors = 0,
  number_of_variables = 3,number_vars_estimated = 3, num_factors_may_vary = FALSE)[[1]]
}
