% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{estimate_beta}
\alias{estimate_beta}
\title{Estimates beta.}
\usage{
estimate_beta(
  NN = aantal_N,
  TT = aantal_T,
  number_of_groups = aantalgroepen,
  number_of_group_factors = aantalfactoren_groups,
  number_of_common_factors = aantalfactoren_common,
  number_of_variables = aantalvars,
  number_vars_estimated = number_variables_estimated,
  num_factors_may_vary = aantalfactors_verschillend_per_group,
  use_median_of_individual_beta = exists("USE_MEDIAN_OF_INDIVIDUAL_THETA"),
  optimize_kappa = FALSE,
  eclipz = FALSE
)
}
\arguments{
\item{NN}{number of individuals}

\item{TT}{length of time series}

\item{number_of_groups}{number of groups estimated}

\item{number_of_group_factors}{number of groupfactors estimated}

\item{number_of_common_factors}{number of common factors estimated}

\item{number_of_variables}{number of observable variables}

\item{number_vars_estimated}{number of variables that are included in the algorithm and have their coefficient estimated. This is usually equal to number_of_variables.}

\item{num_factors_may_vary}{whether or not the number of groupfactors is constant over all groups or not}

\item{use_median_of_individual_beta}{for testing purposes}

\item{optimize_kappa}{indicates if kappa has to be optimized or not; defaults to FALSE}

\item{eclipz}{parameter to indicate using real world Eclipzdataset. Defaults to FALSE.}
}
\value{
list: 1st element contains matrix (N columns: 1 for each element of the panel data) with estimated beta_est's.
}
\description{
Update step of algorithm to obtain new estimation for beta. Note that we call it beta_est because beta() exists in base R.
}
\examples{
#This function needs several initial parameters to be initialized in order to work on itself.
library(RCTS)
library(tidyverse)
library(robustbase)

X = RCTS::X_dgp3
Y = RCTS::Y_dgp3
#Set estimations for group factors and its loadings, and group membership to the real value
lambda_group = RCTS::lambda_group_real_dgp3
factor_group = RCTS::factor_group_real_dgp3
g = RCTS::g_real_dgp3
#There are no common factors to be estimated  -> but needs placeholder
lambda = matrix(0,nrow=1,ncol=300)
comfactor = matrix(0,nrow=1,ncol=30)
use_robust = TRUE
#Choose how coefficients of the observable variables are estimated
homogeneous_coefficients = FALSE
heterogeneous_coefficients_groups = FALSE
heterogeneous_coefficients_individuals = TRUE
ABDGP1 = FALSE
ABintercept = TRUE
use_macropca_instead_of_cz = TRUE
beta_est = estimate_beta(NN = 300, TT = 30,
  number_of_groups = 3, number_of_group_factors = c(3,3,3), number_of_common_factors = 0,
 number_of_variables = 3,number_vars_estimated=3,num_factors_may_vary=FALSE)[[1]]
}
