% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{estimate_factor}
\alias{estimate_factor}
\title{Estimates common factor F}
\usage{
estimate_factor(
  theta,
  g,
  lgfg_list,
  initialise = FALSE,
  AANTALFACTOREN_COMMON = aantalfactoren_common
)
}
\arguments{
\item{theta}{beta}

\item{g}{vector with group membership}

\item{lgfg_list}{This is a list (length number of groups) containing FgLg for every group.}

\item{initialise}{boolean}

\item{AANTALFACTOREN_COMMON}{number of common factors}
}
\value{
r x T matrix
}
\description{
the estimator for F, see Anderson (1984), is equal to the first r eigenvectors (multiplied by sqrt(T) due to the restriction F'F/T = I)
associated with first r largest eigenvalues of the matrix WW' (which is TxT)
W is called Z in Ando / Bai (2017)
}
