% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{estimate_factor}
\alias{estimate_factor}
\title{Helpfunction in estimate_factor() for the robust approach.}
\usage{
estimate_factor(
  theta,
  g,
  lgfg_list,
  initialise = FALSE,
  NN = aantal_N,
  TT = aantal_T,
  number_of_common_factors = aantalfactoren_common
)
}
\arguments{
\item{theta}{theta}

\item{g}{vector with group membership}

\item{lgfg_list}{This is a list (length number of groups) containing FgLg for every group.}

\item{initialise}{boolean}

\item{NN}{number of individuals}

\item{TT}{length of time series}

\item{number_of_common_factors}{number of common factors estimated}
}
\value{
r x T matrix
}
\description{
It has some options used for testing.
Estimates common factor F
}
\details{
The estimator for F, see Anderson (1984), is equal to the first r eigenvectors (multiplied by sqrt(T) due to the restriction F'F/T = I)
associated with first r largest eigenvalues of the matrix WW' (which is TxT)
W is called Z in Ando/Bai (2017)
}
