% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{estimate_factor_group}
\alias{estimate_factor_group}
\title{Estimates group factors}
\usage{
estimate_factor_group(
  beta_est,
  g,
  lambda,
  comfactor,
  use_macropca_instead_of_cz,
  initialise = FALSE,
  NN = number_of_time_series,
  TT = length_of_time_series,
  number_of_groups = number_of_groups_fixedvalue,
  number_of_group_factors = number_of_group_factors_fixedvalue,
  number_of_common_factors = number_of_common_factors_fixedvalue,
  number_of_variables = number_of_variables_fixedvalue,
  number_vars_estimated = number_vars_estimated_fixedvalue,
  verbose = FALSE
)
}
\arguments{
\item{beta_est}{beta_est}

\item{g}{vector with group membership}

\item{lambda}{common factor loadings}

\item{comfactor}{common factors}

\item{use_macropca_instead_of_cz}{If TRUE, then factors are estimated robustly with macropca (or pertMM), else with class-zero-method.}

\item{initialise}{boolean}

\item{NN}{number of individuals}

\item{TT}{length of time series}

\item{number_of_groups}{number of groups estimated}

\item{number_of_group_factors}{number of groupfactors to be estimated}

\item{number_of_common_factors}{number of common factors to be estimated}

\item{number_of_variables}{number of observable variables}

\item{number_vars_estimated}{number of variables that are included in the algorithm and have their coefficient estimated. This is usually equal to number_of_variables.}

\item{verbose}{when TRUE, it prints messages}
}
\description{
Estimates group factors
}
