% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{estimate_factor_group}
\alias{estimate_factor_group}
\title{Estimates group factors}
\usage{
estimate_factor_group(
  beta_est,
  g,
  lambda,
  comfactor,
  use_macropca_instead_of_cz,
  initialise = FALSE,
  NN = aantal_N,
  TT = aantal_T,
  number_of_groups = aantalgroepen,
  number_of_group_factors = aantalfactoren_groups,
  number_of_common_factors = aantalfactoren_common,
  number_of_variables = aantalvars,
  number_vars_estimated = number_variables_estimated,
  eclipz = eclipz,
  expert_based_initial_factors = FALSE
)
}
\arguments{
\item{beta_est}{beta_est}

\item{g}{vector with group membership}

\item{lambda}{common factor loadings}

\item{comfactor}{common factors}

\item{use_macropca_instead_of_cz}{If TRUE, then factors are estimated robustly with macropca (or pertMM), else with class-zero-method.}

\item{initialise}{boolean}

\item{NN}{number of individuals}

\item{TT}{length of time series}

\item{number_of_groups}{number of groups estimated}

\item{number_of_group_factors}{number of groupfactors to be estimated}

\item{number_of_common_factors}{number of common factors to be estimated}

\item{number_of_variables}{number of observable variables}

\item{number_vars_estimated}{number of variables that are included in the algorithm and have their coefficient estimated. This is usually equal to number_of_variables.}

\item{eclipz}{parameter to indicate using real world Eclipzdataset. Defaults to FALSE.}

\item{expert_based_initial_factors}{indicates using a previously defined set of (initial) factors; defaults to FALSE}
}
\description{
Estimates group factors
}
