% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{estimate_factor_group}
\alias{estimate_factor_group}
\title{Estimates group factors}
\usage{
estimate_factor_group(
  use_robust,
  Y,
  X,
  beta_est,
  g,
  lambda,
  comfactor,
  method_estimate_factors,
  method_estimate_beta,
  number_of_groups,
  number_of_group_factors,
  number_of_common_factors,
  number_of_variables,
  number_vars_estimated,
  initialise = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{use_robust}{TRUE or FALSE: defines using the classical or robust algorithm to estimate beta}

\item{Y}{Y}

\item{X}{X}

\item{beta_est}{estimated values of beta}

\item{g}{vector with group membership}

\item{lambda}{common factor loadings}

\item{comfactor}{common factors}

\item{method_estimate_factors}{defines method of robust estimaton of the factors: "macro", "pertmm" or "cz"}

\item{method_estimate_beta}{defines how beta is estimated. Default case is an estimated beta for each individual. Default value is "individual." Possible values are "homogeneous", "group" or "individual".}

\item{number_of_groups}{number of groups estimated}

\item{number_of_group_factors}{number of groupfactors to be estimated}

\item{number_of_common_factors}{number of common factors to be estimated}

\item{number_of_variables}{number of observable variables}

\item{number_vars_estimated}{number of variables that are included in the algorithm and have their coefficient estimated. This is usually equal to number_of_variables.}

\item{initialise}{boolean}

\item{verbose}{when TRUE, it prints messages}
}
\description{
Estimates group factors
}
