% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{estimate_factor_group}
\alias{estimate_factor_group}
\title{estimates group factors}
\usage{
estimate_factor_group(
  theta,
  g,
  lambda,
  comfactor,
  initialise = FALSE,
  AANTALFACTOREN_GROUPS = aantalfactoren_groups
)
}
\arguments{
\item{theta}{beta}

\item{g}{vector with group membership}

\item{lambda}{loadings}

\item{comfactor}{common factors}

\item{initialise}{boolean}

\item{AANTALFACTOREN_GROUPS}{vector with number of group factors}
}
\description{
estimates group factors
}
