% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{evade_crashes_macropca}
\alias{evade_crashes_macropca}
\title{Solves a very specific issue with MacroPCA.}
\usage{
evade_crashes_macropca(object, verbose = FALSE)
}
\arguments{
\item{object}{input}

\item{verbose}{prints messages}
}
\description{
MacroPCA crashes Rstudio with certain dimensions of the input. Solve this by doubling every row.
No information is added by this, so there is no influence on the end result,
but crashes of Rstudio are evaded.
}
