% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{fill_rc}
\alias{fill_rc}
\title{Fills in the optimized number of common factors for each C.}
\usage{
fill_rc(df, all_best_values, subset)
}
\arguments{
\item{df}{input}

\item{all_best_values}{data frame with the optimal number of groups, common factors and group specific factors}

\item{subset}{index of the subsample}
}
\description{
Fills in the optimized number of common factors for each C.
}
