% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{fill_rcj}
\alias{fill_rcj}
\title{Fills in the optimized number of groups and group specific factors for each C.}
\usage{
fill_rcj(df, all_best_values, subset, S_cand, kg_cand)
}
\arguments{
\item{df}{input}

\item{all_best_values}{data frame with the optimal number of groups, common factors and group specific factors}

\item{subset}{index of the subsample}

\item{S_cand}{candidate values for the number of estimated groups}

\item{kg_cand}{candidate values for the number of estimated group specific factors}
}
\description{
Fills in the optimized number of groups and group specific factors for each C.
}
