% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{generate_grouped_factorstructure}
\alias{generate_grouped_factorstructure}
\title{Generates the real groupfactorstructure, to use in simulations.}
\usage{
generate_grouped_factorstructure(S, aantalfactoren_groups_real)
}
\arguments{
\item{S}{real number of groups}

\item{aantalfactoren_groups_real}{vector with as length the number of groups, where each element is the number of real groupfactors of that group.}
}
\value{
list: first element contains real groupfactors and second element contains real groupfactor loadings
}
\description{
Loadings and factors are generated by:
loadings ~ N(LGR_FACTOR_mean, j * LGR_FACTOR) -> default case will be N(0,j)
factors ~ N(j * FGR_FACTOR, FGR_FACTOR_sd) -> default case will be N(j,1)
}
\examples{
#3 groups, each with 3 groupfactors:
generate_grouped_factorstructure(3,c(3,3,3))
}
