% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{generate_grouped_factorstructure}
\alias{generate_grouped_factorstructure}
\title{Generates the true groupfactorstructure, to use in simulations.}
\usage{
generate_grouped_factorstructure(
  S,
  true_number_of_group_factors,
  TT = length_of_time_series_fulldata,
  g_true
)
}
\arguments{
\item{S}{true number of groups}

\item{true_number_of_group_factors}{vector with as length the number of groups, where each element is the true number of groupfactors of that group.}

\item{TT}{length of time series}

\item{g_true}{true group membership}
}
\value{
list: first element contains true groupfactors and second element contains true groupfactor loadings
}
\description{
Loadings and factors are generated by:
loadings ~ N(LGR_FACTOR_mean, j * LGR_FACTOR) -> default case will be N(0,j)
factors ~ N(j * FGR_FACTOR, FGR_FACTOR_sd) -> default case will be N(j,1)
}
\examples{
library(tidyverse)
#For 3 groups, each with 3 groupfactors:
g_true = ceiling(runif(300) * 3)
generate_grouped_factorstructure(3, c(3, 3, 3), TT = 30, g_true)
}
