% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{generate_Y}
\alias{generate_Y}
\title{Generate panel data Y for simulations.}
\usage{
generate_Y(
  NN,
  TT,
  true_number_of_common_factors,
  true_number_of_group_factors,
  g_true,
  beta_true,
  lambda_group_true,
  factor_group_true,
  lambda_true,
  comfactor_true,
  epsilon,
  X,
  number_of_variables = number_of_variables_fixedvalue
)
}
\arguments{
\item{NN}{number of individuals}

\item{TT}{length of time series}

\item{true_number_of_common_factors}{true number of common factors}

\item{true_number_of_group_factors}{Vector of length the number of groups. Each element contains the true number of group factors for that group.}

\item{g_true}{vector of length N with true group memberships}

\item{beta_true}{true coefficients with the observable variables}

\item{lambda_group_true}{loadings of the group factors}

\item{factor_group_true}{groupfactors}

\item{lambda_true}{loadings of the common factors}

\item{comfactor_true}{common factors}

\item{epsilon}{NN x TT-matrix containing the error term}

\item{X}{dataframe with the observed variables}

\item{number_of_variables}{number of observable variables}
}
\value{
N x T matrix
}
\description{
Generate panel data Y for simulations.
}
