% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{generate_Y}
\alias{generate_Y}
\title{Generate panel data Y for simulations.}
\usage{
generate_Y(
  NN,
  TT,
  number_of_common_factors_real,
  number_of_group_factors_real,
  g_real,
  theta_real,
  lambda_group_real,
  factor_group_real,
  lambda_real,
  factor_real,
  epsilon,
  abctypes = FALSE,
  ando_bai = FALSE,
  ando_bai_2017 = FALSE,
  eclipz = FALSE,
  number_of_Variables = aantalvars
)
}
\arguments{
\item{NN}{N: number of elements}

\item{TT}{T: length of timer series}

\item{number_of_common_factors_real}{real number of common factors}

\item{number_of_group_factors_real}{Vector of length the number of groups. Each element contains the real number of group factors for that group.}

\item{g_real}{vector with real group memberships}

\item{theta_real}{real coefficients with the observable variables}

\item{lambda_group_real}{loadings of the group factors}

\item{factor_group_real}{groupfactors}

\item{lambda_real}{loadings of the common factors}

\item{factor_real}{common factors}

\item{epsilon}{NN x TT-matrix containing the error term}

\item{abctypes}{option to make a dataset containing specific incomepatterns based on incomelimits for Belgian independents. Only for testing purposes. Defaults to FALSE.}

\item{ando_bai}{loads Ando/Bai-dataset; only for testing purposes. Defaults to FALSE.}

\item{ando_bai_2017}{loads Ando/Bai-dataset; only for testing purposes. Defaults to FALSE.}

\item{eclipz}{parameter to indicate using real world Eclipzdataset. Defaults to FALSE.}

\item{number_of_Variables}{number of observable variables}
}
\value{
N x T matrix
}
\description{
Generate panel data Y for simulations.
}
