% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{get_convergence_speed}
\alias{get_convergence_speed}
\title{This function defines the convergence speed.}
\usage{
get_convergence_speed(
  iteration,
  use_robust,
  of,
  length_of_time_series = nrow(Y),
  number_of_time_series = ncol(Y),
  verbose = FALSE
)
}
\description{
This function defines the convergence speed.
}
