% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{get_final_estimation}
\alias{get_final_estimation}
\title{Function that returns the final clustering, based on the estimated number of groups and common and group specific factors.}
\usage{
get_final_estimation(df, opt_groups, k, kg, type, limit_est_groups = 20)
}
\arguments{
\item{df}{input dataframe (this will be df_results_full)}

\item{opt_groups}{the optimal number of groups}

\item{k}{the optimal number of common factors}

\item{kg}{vector with the optimal number of group specific factors}

\item{type}{defines which estimation to return: options are "clustering", "beta", "fg" (group specific factors), "lg" (loadings corresponding to fg), "f" (common factors), "l" (loadings corresponding to f),}

\item{limit_est_groups}{maximum allowed number of groups that can be estimated}
}
\description{
Function that returns the final clustering, based on the estimated number of groups and common and group specific factors.
}
