% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{grid_add_variables}
\alias{grid_add_variables}
\title{Function which is used to have a df ("grid") with data (individualindex, timeindex, XT and LF) available.}
\usage{
grid_add_variables(
  grid,
  theta,
  lambda,
  comfactor,
  number_of_variables = aantalvars,
  number_of_groups = aantalgroepen
)
}
\arguments{
\item{grid}{dataframe containing values for X*beta and LF (product of common factor and its loadings)}

\item{theta}{beta}

\item{lambda}{loadings of the common factors}

\item{comfactor}{common factors}

\item{number_of_variables}{number of observable variables}

\item{number_of_groups}{number of groups estimated}
}
\description{
It is used in update_g().
}
