% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{grid_add_variables}
\alias{grid_add_variables}
\title{Function which is used to have a dataframe (called "grid") with data (individualindex, timeindex, XT and LF) available.}
\usage{
grid_add_variables(
  grid,
  beta_est,
  lambda,
  comfactor,
  method_estimate_beta,
  NN = number_of_time_series,
  TT = length_of_time_series,
  number_of_variables = number_of_variables_fixedvalue,
  number_vars_estimated = number_vars_estimated_fixedvalue,
  number_of_groups = number_of_groups_fixedvalue,
  limit_est_groups_heterogroups = 15
)
}
\arguments{
\item{grid}{dataframe containing values for X*beta_est and LF (product of common factor and its loadings)}

\item{beta_est}{estimated values of beta}

\item{lambda}{loadings of the common factors}

\item{comfactor}{common factors}

\item{method_estimate_beta}{defines how beta is estimated. Default case is an estimated beta for each individual. Default value is "individual." Possible values are "homogeneous", "group" or "individual".}

\item{NN}{number of individuals}

\item{TT}{length of time series}

\item{number_of_variables}{number of observable variables}

\item{number_vars_estimated}{number of variables that are included in the algorithm and have their coefficient estimated. This is usually equal to number_of_variables.}

\item{number_of_groups}{number of groups estimated}

\item{limit_est_groups_heterogroups}{maximum amount of groups that can be estimated when method_estimate_beta is set to "group"}
}
\description{
(The dataframe "grid" used to be required in update_g(), via OF_vectorized3(), but not anymore.)
}
