% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{grid_add_variables}
\alias{grid_add_variables}
\title{Function which is used to have a df ("grid") with data (individualindex, timeindex, XT and LF) available.}
\usage{
grid_add_variables(grid, theta, lambda, comfactor)
}
\arguments{
\item{theta}{beta}

\item{lambda}{loadings of the common factors}

\item{comfactor}{common factors}
}
\description{
It is used in update_g().
}
