% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{grid_add_variables}
\alias{grid_add_variables}
\title{Function which is used to have a dataframe (called "grid") with data (individualindex, timeindex, XT and LF) available.}
\usage{
grid_add_variables(
  grid,
  beta_est,
  lambda,
  comfactor,
  NN = number_of_time_series,
  TT = length_of_time_series,
  number_of_variables = number_of_variables_fixedvalue,
  number_vars_estimated = number_vars_estimated_fixedvalue,
  number_of_groups = number_of_groups_fixedvalue
)
}
\arguments{
\item{grid}{dataframe containing values for X*beta_est and LF (product of common factor and its loadings)}

\item{beta_est}{estimated values of beta}

\item{lambda}{loadings of the common factors}

\item{comfactor}{common factors}

\item{NN}{number of individuals}

\item{TT}{length of time series}

\item{number_of_variables}{number of observable variables}

\item{number_vars_estimated}{number of variables that are included in the algorithm and have their coefficient estimated. This is usually equal to number_of_variables.}

\item{number_of_groups}{number of groups estimated}
}
\description{
The dataframe "grid" is required in update_g(), via OF_vectorized3().
}
