% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{handle_macropca_errors}
\alias{handle_macropca_errors}
\title{Helpfunction in bepaal_robpca().}
\usage{
handle_macropca_errors(object, temp, KMAX, aantal_eigenvectoren)
}
\description{
It handles possible thrown errors in MacroPCA.
}
