% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_cleaning.R
\name{handleNA_LG}
\alias{handleNA_LG}
\title{Removes NA's in LG (in function calculate_virtual_factor_and_lambda_group() )}
\usage{
handleNA_LG(df)
}
\arguments{
\item{df}{input}
}
\value{
matrix
}
\description{
Removes NA's in LG (in function calculate_virtual_factor_and_lambda_group() )
}
