% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{initialise_beta}
\alias{initialise_beta}
\title{Initialisation of estimation of beta.}
\usage{
initialise_beta(
  eclipz = FALSE,
  NN = aantal_N,
  TT = aantal_T,
  number_of_Variables = aantalvars,
  number_vars_estimated = SCHATTEN_MET_AANTALVARS,
  number_of_groups = aantalgroepen
)
}
\arguments{
\item{eclipz}{parameter to indicate using real world Eclipzdataset. Defaults to FALSE.}

\item{NN}{number of individuals}

\item{TT}{length of time series}
}
\description{
Note: this needs to be called before the definition of grid.
}
\examples{
library(RobClustTimeSeries)
library(tidyverse)
library(robustbase)
lambda_group = RobClustTimeSeries::lambda_group_real_dgp3
factor_group = RobClustTimeSeries::factor_group_real_dgp3
lambda = matrix(0,nrow=1,ncol=300) #0 factors -> but needs placeholder
comfactor = matrix(0,nrow=1,ncol=30) #0 factors -> but needs placeholder
X = RobClustTimeSeries::X_dgp3
Y = RobClustTimeSeries::Y_dgp3
g = RobClustTimeSeries::g_real_dgp3

homogeneous_coefficients = FALSE
heterogeneous_coefficients_groups = FALSE
heterogeneous_coefficients_individuals = TRUE #estimating beta_i for every individual
use_robust = TRUE
ABDGP1 = FALSE
ABintercept = TRUE
beta_init = initialise_beta(NN = 300, TT = 30, number_of_Variables = 3, number_vars_estimated = 3, number_of_groups = 3)
}
