% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{initialise_beta}
\alias{initialise_beta}
\title{Initialisation of estimation of beta (the coefficients with the observable variables)}
\usage{
initialise_beta(
  robust,
  Y,
  X,
  S,
  g,
  method_estimate_beta = "individual",
  nosetting_lmrob = FALSE,
  special_case_dgp1 = FALSE
)
}
\arguments{
\item{robust}{robust or classical estimation}

\item{Y}{Y: NxT dataframe with the panel data of interest}

\item{X}{dataframe with the observed variables}

\item{S}{estimated number of groups}

\item{g}{vector with estimated group membership for all individuals}

\item{method_estimate_beta}{defines how beta is estimated. Default case is an estimated beta for each individual. Default value is "individual." Possible values are "homogeneous", "group" or "individual".}

\item{nosetting_lmrob}{option to remove the recommended setting in lmrob(). It is much faster. Defaults to FALSE.}

\item{special_case_dgp1}{special case for data generated according to dgp 1: it changes the 1st variable in X to 1 (-> intercept). Consequently the estimation of beta needs to be restructured slightly.}
}
\description{
Note: this needs to be called before the definition of grid.
}
\examples{
\donttest{
X <- RCTS::X_dgp3
Y <- RCTS::Y_dgp3
# Set estimations for group factors and its loadings, and group membership
#  to the true value for this example.
lambda_group <- RCTS::lambda_group_true_dgp3
factor_group <- RCTS::factor_group_true_dgp3
g <- RCTS::g_true_dgp3
beta_init <- initialise_beta(TRUE, Y, X,
  S = 3, g
)
}
}
