% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{initialise_beta}
\alias{initialise_beta}
\title{Initialisation of estimation of beta (the coefficients with the observable variables)}
\usage{
initialise_beta(
  NN,
  TT,
  number_of_variables,
  number_vars_estimated,
  number_of_groups,
  method_estimate_beta = "individual"
)
}
\arguments{
\item{NN}{number of individuals}

\item{TT}{length of time series}

\item{number_of_variables}{number of observable variables}

\item{number_vars_estimated}{number of variables from which the coefficients are estimated}

\item{number_of_groups}{number of groups}

\item{method_estimate_beta}{"individual" #can be "homogeneous", "group" or "individual"}
}
\description{
Note: this needs to be called before the definition of grid.
}
\examples{
library(RCTS)
library(tidyverse)
library(robustbase)

X = RCTS::X_dgp3
Y = RCTS::Y_dgp3
#Set estimations for group factors and its loadings, and group membership to the true value
lambda_group = RCTS::lambda_group_true_dgp3
factor_group = RCTS::factor_group_true_dgp3
g = RCTS::g_true_dgp3
#There are no common factors to be estimated  -> but needs placeholder
lambda = matrix(0,nrow=1,ncol=300)
comfactor = matrix(0,nrow=1,ncol=30)

use_robust = TRUE
#Choose how coefficients of the observable are estimated
method_estimate_beta = "individual" #estimating beta_i for every individual
dgp1_AB = FALSE
dgp1_spread_group_centers = TRUE #this value is only relevant if dgp1_AB ise set to TRUE
beta_init = initialise_beta(NN = 300, TT = 30,
  number_of_variables = 3, number_vars_estimated = 3, number_of_groups = 3)
}
