% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{initialise_clustering}
\alias{initialise_clustering}
\title{Function that clusters time series in a dataframe with kmeans (classical algorithm) or trimmed kmeans(robust algorithms).}
\usage{
initialise_clustering(
  robust,
  Y,
  g,
  beta_est,
  S,
  k,
  kg,
  comfactor,
  max_percent_outliers_tkmeans = 0,
  verbose = FALSE
)
}
\arguments{
\item{robust}{TRUE or FALSE: defines using the classical or robust algorithm to estimate beta}

\item{Y}{Y: NxT dataframe with the panel data of interest}

\item{g}{Vector with estimated group membership for all individuals}

\item{beta_est}{estimated values of beta}

\item{S}{the desired number of groups}

\item{k}{number of common factors to be estimated}

\item{kg}{number of group specific factors to be estimated}

\item{comfactor}{estimated common factors}

\item{max_percent_outliers_tkmeans}{The proportion of observations to be trimmed.}

\item{verbose}{when TRUE, it prints messages}
}
\description{
If a time series contains NA's a random cluster will be assigned to that time series.
}
