% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{initialise_commonfactorstructure_macropca}
\alias{initialise_commonfactorstructure_macropca}
\title{Initialises the estimation of the common factors and their loadings.}
\usage{
initialise_commonfactorstructure_macropca(
  robust,
  Y,
  X,
  beta_est,
  g,
  factor_group,
  k,
  kg,
  method_estimate_beta = "individual",
  method_estimate_factors = "macro",
  vars_est = dim(X)[3],
  verbose = FALSE
)
}
\arguments{
\item{robust}{robust or classical estimation}

\item{Y}{Y: NxT dataframe with the panel data of interest}

\item{X}{dataframe with the observed variables}

\item{beta_est}{estimated values of beta}

\item{g}{Vector with estimated group membership for all individuals}

\item{factor_group}{estimated group specific factors}

\item{k}{number of estimated common factors}

\item{kg}{vector with the number of estimated group specific factors}

\item{method_estimate_beta}{defines how beta is estimated. Default case is an estimated beta for each individual. Default value is "individual." Possible values are "homogeneous", "group" or "individual".}

\item{method_estimate_factors}{specifies the robust algorithm to estimate factors: default is "macro". The value is not used when robust is set to FALSE.}

\item{vars_est}{number of available observed variables for which a coefficient will be estimated. As default it is equal to the number of available observed variables.}

\item{verbose}{when TRUE, it prints messages}
}
\description{
This is a short version of initialise_commonfactorstructure() which only contains implementations for the robust macropca case and the classical case.
}
