% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{initialise_df_pic}
\alias{initialise_df_pic}
\title{Initialises a dataframe which will contain the PIC for each configuration and for each value of C.}
\usage{
initialise_df_pic(C_candidates)
}
\arguments{
\item{C_candidates}{candidates for C (parameter in PIC)}
}
\value{
Returns an empty data.frame.
}
\description{
Initialises a dataframe which will contain the PIC for each configuration and for each value of C.
}
\examples{
initialise_df_pic(1:10)
}
