% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{initialise_df_results}
\alias{initialise_df_results}
\title{Initialises a dataframe that will contain an overview of metrics for each estimated configuration (for example adjusted randindex).}
\usage{
initialise_df_results(robust, limit_est_groups = 20)
}
\arguments{
\item{robust}{robust or classical estimation}

\item{limit_est_groups}{maximum allowed number of groups that can be estimated}
}
\value{
Returns an empty data.frame.
}
\description{
Initialises a dataframe that will contain an overview of metrics for each estimated configuration (for example adjusted randindex).
}
\examples{
initialise_df_results(TRUE)
}
