% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{initialise_theta}
\alias{initialise_theta}
\title{Initialisation of estimation of beta (the coefficients with the observable variables)}
\usage{
initialise_theta(
  eclipz = FALSE,
  NN = aantal_N,
  TT = aantal_T,
  number_of_variables = aantalvars,
  number_vars_estimated = SCHATTEN_MET_AANTALVARS,
  number_of_groups = aantalgroepen
)
}
\arguments{
\item{eclipz}{parameter to indicate using real world Eclipzdataset. Defaults to FALSE.}

\item{NN}{number of individuals}

\item{TT}{length of time series}

\item{number_of_variables}{number of observable variables}

\item{number_vars_estimated}{number of variables from which the coefficients are estimated}

\item{number_of_groups}{number of groups}
}
\description{
Note: this needs to be called before the definition of grid.
}
\examples{
library(RobClustTimeSeries)
library(tidyverse)
library(robustbase)

X = RobClustTimeSeries::X_dgp3
Y = RobClustTimeSeries::Y_dgp3
#Set estimations for group factors and its loadings, and group membership to the real value
lambda_group = RobClustTimeSeries::lambda_group_real_dgp3
factor_group = RobClustTimeSeries::factor_group_real_dgp3
g = RobClustTimeSeries::g_real_dgp3
#There are no common factors to be estimated  -> but needs placeholder
lambda = matrix(0,nrow=1,ncol=300)
comfactor = matrix(0,nrow=1,ncol=30)

use_robust = TRUE
#Choose how coefficients of the observable are estimated
homogeneous_coefficients = FALSE
heterogeneous_coefficients_groups = FALSE
heterogeneous_coefficients_individuals = TRUE #estimating theta_i for every individual
ABDGP1 = FALSE
ABintercept = TRUE
theta_init = initialise_theta(NN = 300, TT = 30,
  number_of_variables = 3, number_vars_estimated = 3, number_of_groups = 3)
}
