% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{iterate}
\alias{iterate}
\title{Wrapper around estimate_beta, update_g, and estimating the factorstructures.}
\usage{
iterate(
  use_robust,
  Y,
  X,
  beta_est,
  g,
  lambda_group,
  factor_group,
  lambda,
  comfactor,
  S,
  k,
  kg,
  method_estimate_beta = "individual",
  method_estimate_factors = "macro",
  vars_est = 3,
  special_case_dgp1 = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{use_robust}{TRUE or FALSE: defines using the classical or robust algorithm to estimate beta}

\item{Y}{Y: the panel data of interest}

\item{X}{X}

\item{beta_est}{estimated values of beta}

\item{g}{Vector with estimated group membership for all individuals}

\item{lambda_group}{loadings of the estimated group specific factors}

\item{factor_group}{estimated group specific factors}

\item{lambda}{loadings of the estimated common factors}

\item{comfactor}{estimated common factors}

\item{S}{number of groups to estimate}

\item{k}{number of common factors to estimate}

\item{kg}{vector with length S. Each element contains the number of group specific factors to estimate.}

\item{method_estimate_beta}{defines how beta is estimated. Default case is an estimated beta for each individual. Default value is "individual." Possible values are "homogeneous", "group" or "individual".}

\item{method_estimate_factors}{specifies the robust algorithm to estimate factors: default is "macro". The value is not used when use_robust is set to FALSE.}

\item{vars_est}{number of variables for which a beta is estimated. Usually equal to the number of variables available in X.}

\item{special_case_dgp1}{TRUE or FALSE: whether data is generated from dgp1 and has the extra spread in group centers. Default is FALSE.}

\item{verbose}{when TRUE, it prints messages}
}
\description{
Wrapper around estimate_beta, update_g, and estimating the factorstructures.
}
