% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{LMROB}
\alias{LMROB}
\title{Wrapper around lmrob, to make sure this error does not happen anymore.}
\usage{
LMROB(parameter_y, parameter_x, NOINTERCEPT = FALSE)
}
\arguments{
\item{parameter_y}{dependent variable}

\item{parameter_x}{independent variables}

\item{NOINTERCEPT}{if TRUE it performs regression without an intercept}
}
\description{

}
