% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_parallel.R
\name{make_df_pic_parallel}
\alias{make_df_pic_parallel}
\title{Makes a dataframe with the PIC for each configuration and each candidate C.}
\usage{
make_df_pic_parallel(x, C_candidates)
}
\arguments{
\item{x}{output of the parallel version of the algorithm}

\item{C_candidates}{candidates for C}
}
\value{
data.frame
}
\description{
Makes a dataframe with the PIC for each configuration and each candidate C.
}
