% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{make_subsamples}
\alias{make_subsamples}
\title{This function selects a subsample of the time series, and of the length of the time series. Based on this it returns a subsample of Y.
It also returns the corresponding subsamples X and the group membership.}
\usage{
make_subsamples(
  Y,
  X,
  number_of_time_series_fulldata,
  length_of_time_series_fulldata,
  stepsize,
  stepsize_N = round(number_of_time_series_fulldata/10),
  stepsize_T = round(length_of_time_series_fulldata/30)
)
}
\arguments{
\item{Y}{input Y}

\item{X}{input X}

\item{number_of_time_series_fulldata}{number of time series of the original dataset}

\item{length_of_time_series_fulldata}{length of time series of the original dataset}

\item{stepsize}{index of the subsample: this defines how many times stepsize_N is subtracted from the original N time series. Similar for stepsize_T.}

\item{stepsize_N}{size of the decrease in N; multiplied with stepsize}

\item{stepsize_T}{size of the decrease in T; multiplied with stepsize}
}
\description{
This function selects a subsample of the time series, and of the length of the time series. Based on this it returns a subsample of Y.
It also returns the corresponding subsamples X and the group membership.
}
