% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{mse_heterogeneous_groups}
\alias{mse_heterogeneous_groups}
\title{Helpfunction in calculate_mse_beta, when method_estimate_beta == "group".
(beta is estimated for each group separately).}
\usage{
mse_heterogeneous_groups(without_intercept, dgp1_spread_group_centers)
}
\arguments{
\item{without_intercept}{boolean to remove the intercept in the calculation}

\item{dgp1_spread_group_centers}{boolean to remove the de facto intercept in the calculation (this is specific for DGP 1, due to how this is defined.)}
}
\description{
Helpfunction in calculate_mse_beta, when method_estimate_beta == "group".
(beta is estimated for each group separately).
}
