% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{OF_vectorized3}
\alias{OF_vectorized3}
\title{Calculates objective function: used in local_search + to determine "best_result".}
\usage{
OF_vectorized3(
  group_memberships,
  THETA = theta,
  ALPHA = alpha,
  LAMBDA = lambda,
  FACTOR = comfactor,
  LAMBDA_GROUP = lambda_group,
  FACTOR_GROUP = factor_group
)
}
\arguments{
\item{group_memberships}{Vector containing groupmembership for all individuals.}

\item{THETA}{beta}

\item{ALPHA}{--obsolete--}

\item{LAMBDA}{loadings}

\item{FACTOR}{common factors}

\item{LAMBDA_GROUP}{grouploadings}

\item{FACTOR_GROUP}{groupfactors}
}
\description{
Calculates objective function: used in local_search + to determine "best_result".
}
