% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{OF_vectorized3}
\alias{OF_vectorized3}
\title{Calculates objective function: used in local_search + to determine "best_result".}
\usage{
OF_vectorized3(
  group_memberships,
  beta_est = beta_est,
  lc = lambda,
  fc = comfactor,
  lg = lambda_group,
  fg = factor_group,
  NN = number_of_time_series,
  number_of_groups = number_of_groups_fixedvalue,
  number_of_common_factors = number_of_common_factors_fixedvalue,
  number_of_group_factors = number_of_group_factors_fixedvalue,
  num_factors_may_vary = num_factors_may_vary_fixedvalue
)
}
\arguments{
\item{group_memberships}{Vector containing the group membership for all individuals.}

\item{beta_est}{estimated values of beta}

\item{lc}{loadings of estimated common factors}

\item{fc}{estimated common factors}

\item{lg}{estimated grouploadings}

\item{fg}{estimated groupfactors}

\item{NN}{number of individuals}

\item{number_of_groups}{number of groups estimated}

\item{number_of_common_factors}{number of common factors to be estimated}

\item{number_of_group_factors}{number of groupfactors to be estimated}

\item{num_factors_may_vary}{whether or not the number of groupfactors is constant over all groups or not}
}
\description{
Calculates objective function: used in local_search + to determine "best_result".
}
