% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{OF_vectorized3}
\alias{OF_vectorized3}
\title{Calculates objective function: used in local_search + to determine "best_result".}
\usage{
OF_vectorized3(
  group_memberships,
  THETA = theta,
  ALPHA = alpha,
  LAMBDA = lambda,
  FACTOR = comfactor,
  LAMBDA_GROUP = lambda_group,
  FACTOR_GROUP = factor_group,
  NN = aantal_N,
  number_of_groups = aantalgroepen,
  number_of_common_factors = aantalfactoren_common,
  number_of_group_factors = aantalfactoren_groups,
  num_factors_may_vary = aantalfactors_verschillend_per_group
)
}
\arguments{
\item{group_memberships}{Vector containing groupmembership for all individuals.}

\item{THETA}{beta}

\item{ALPHA}{--obsolete--}

\item{LAMBDA}{loadings}

\item{FACTOR}{common factors}

\item{LAMBDA_GROUP}{grouploadings}

\item{FACTOR_GROUP}{groupfactors}

\item{NN}{number of individuals}

\item{number_of_groups}{number of groups estimated}

\item{number_of_common_factors}{number of common factors estimated}

\item{number_of_group_factors}{number of groupfactors estimated}

\item{num_factors_may_vary}{whether or not the number of groupfactors is constant over all groups or not}
}
\description{
Calculates objective function: used in local_search + to determine "best_result".
}
