% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{OF_vectorized_helpfunction3}
\alias{OF_vectorized_helpfunction3}
\title{Helpfunction in OF_vectorized3()}
\usage{
OF_vectorized_helpfunction3(i, t, XTHETA, LF, group_memberships, lgfg_list)
}
\description{
Helpfunction in OF_vectorized3()
}
