% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_parallel.R
\name{parallel_algorithm}
\alias{parallel_algorithm}
\title{Wrapper of the loop over the subsets which in turn use the parallelised algorithm.}
\usage{
parallel_algorithm(
  original_data,
  indices_subset,
  S_cand,
  k_cand,
  kg_cand,
  C_candidates,
  robust = TRUE,
  USE_DO = FALSE
)
}
\arguments{
\item{original_data}{list containing the original data (1: Y, 2: X, 3: true group membership)}

\item{indices_subset}{vector with indices of the subsets; starts with zero}

\item{S_cand}{candidates for S (number of groups)}

\item{k_cand}{candidates for k (number of common factors)}

\item{kg_cand}{candidates for kg (number of group specific factors)}

\item{C_candidates}{candidates for C}

\item{robust}{robust or classical estimation}

\item{USE_DO}{if TRUE, then a serialized version is performed ("do" instead of "dopar") (for testing purposes)}
}
\description{
Wrapper of the loop over the subsets which in turn use the parallelised algorithm.
}
