% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{plot_VCsquared}
\alias{plot_VCsquared}
\title{Plots expression(VC^2) along with the corresponding number of groups (orange), common factors (darkblue) and group factors of the first group (lightblue).}
\usage{
plot_VCsquared(
  VC_squared,
  rc,
  rcj,
  C_candidates,
  S_cand,
  xlim_min = 0.001,
  xlim_max = 100,
  add_true_lines = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{VC_squared}{measure of variability in the optimal configuration between the subsets}

\item{rc}{dataframe containg the numer of common factors for all candidate C's and all subsamples}

\item{rcj}{dataframe containg the numer of groupfactors for all candidate C's and all subsamples}

\item{C_candidates}{candidates for C (parameter in PIC)}

\item{S_cand}{candidate numbers for the number of groups}

\item{xlim_min}{starting point of the plot}

\item{xlim_max}{end point of the plot}

\item{add_true_lines}{if set to TRUE, for each C the true number of groups, common factors, and group specific factors of group 1 will be added to the plot}

\item{verbose}{if TRUE, more details are printed}
}
\description{
Plots expression(VC^2) along with the corresponding number of groups (orange), common factors (darkblue) and group factors of the first group (lightblue).
}
