% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{restructure_X_to_order_slowN_fastT}
\alias{restructure_X_to_order_slowN_fastT}
\title{This function restructures X (which is an 3D-array of dimensions (N,T,p) to a 2D-matrix of dimension (NxT,p).}
\usage{
restructure_X_to_order_slowN_fastT(
  X,
  number_of_variables,
  number_vars_estimated
)
}
\arguments{
\item{X}{input}

\item{number_of_variables}{number of observable variables}

\item{number_vars_estimated}{number of variables that are included in the algorithm and have their coefficient estimated. This is usually equal to number_of_variables.}
}
\description{
This 2D-matrix is what Ando/Bai worked with.
}
\examples{
X = initialise_X(300, 30, number_of_variables = 3)
X_restructured = restructure_X_to_order_slowN_fastT(X,
  number_of_variables = 3, number_vars_estimated = 3)
}
