% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{restructure_X_to_order_slowN_fastT}
\alias{restructure_X_to_order_slowN_fastT}
\title{This function restructures X (which is an 3D-array of dimensions (N,T,p) to a 2D-matrix of dimension (NxT,p).}
\usage{
restructure_X_to_order_slowN_fastT(X, eclipz)
}
\arguments{
\item{X}{input}
}
\description{
This 2D-matrix is what Ando/Bai worked with.
}
\examples{
X = initialise_X(300,30)
restructure_X_to_order_slowN_fastT(X, FALSE)

}
