% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{restructure_X_to_order_slowN_fastT}
\alias{restructure_X_to_order_slowN_fastT}
\title{Restructures X (which is an 3D-array of dimensions (N,T,p) to a 2D-matrix of dimension (NxT,p).}
\usage{
restructure_X_to_order_slowN_fastT(X, vars_est)
}
\arguments{
\item{X}{input}

\item{vars_est}{number of variables that will be included in the algorithm and have their coefficient estimated. This is usually equal to the number of observable variables.}
}
\description{
Restructures X (which is an 3D-array of dimensions (N,T,p) to a 2D-matrix of dimension (NxT,p).
}
\examples{
X <- RCTS:::initialise_X(300, 30, vars = 3)
X_restructured <- RCTS:::restructure_X_to_order_slowN_fastT(X,
  vars_est = 3
)
}
