% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_IFE_robust_lambda.R
\name{return_robust_lambdaobject}
\alias{return_robust_lambdaobject}
\title{Calculates robust loadings}
\usage{
return_robust_lambdaobject(
  Y_like_object,
  group,
  type,
  g,
  NN_rrn = number_of_time_series,
  factor_group_rrn = factor_group,
  number_of_group_factors_rrn = number_of_group_factors_fixedvalue,
  comfactor_rrn = comfactor,
  number_of_common_factors_rrn = number_of_common_factors_fixedvalue,
  verbose = FALSE
)
}
\arguments{
\item{Y_like_object}{this is Y_ster or W or W_j}

\item{group}{index of group}

\item{type}{scalar which shows in which setting this function is used}

\item{g}{vector with group memberships}

\item{NN_rrn}{number of time series}

\item{factor_group_rrn}{estimation of groupfactors}

\item{number_of_group_factors_rrn}{number of group factors}

\item{comfactor_rrn}{estimation of common factors}

\item{number_of_common_factors_rrn}{number of common factors}

\item{verbose}{when TRUE, it prints messages}
}
\value{
Nxk dataframe
}
\description{
Uses the almost classical lambda (this is an object of which the mean equals to the classical lambda) to create a robust lambda by using M estimation
}
