% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_IFE_robust_lambda.R
\name{return_robust_lambdaobject}
\alias{return_robust_lambdaobject}
\title{Calculates robust loadings}
\usage{
return_robust_lambdaobject(
  Y_like_object,
  group,
  type,
  NN = aantal_N,
  factor_group_RRN = factor_group,
  number_of_group_factors_RRN = aantalfactoren_groups,
  comfactor_RRN = comfactor,
  number_of_common_factors_RRN = aantalfactoren_common,
  eclipz_RRN = FALSE,
  application_covid = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{Y_like_object}{this is Y_ster or W or W_j}

\item{group}{index of group}

\item{type}{scalar which shows in which setting this function is used}

\item{NN}{N}

\item{factor_group_RRN}{estimation of groupfactors}

\item{number_of_group_factors_RRN}{number of group factors}

\item{comfactor_RRN}{estimation of common factors}

\item{number_of_common_factors_RRN}{number of common factors}

\item{eclipz_RRN}{indicator of using the eclipzdataset}

\item{application_covid}{indicator of using the coviddataset}

\item{verbose}{when TRUE, it prints messages}
}
\value{
Nxk dataframe
}
\description{
Uses the almost classical lambda (this is an object of which the mean equals to the classical lambda) to create a robust lambda by using M estimation
}
