% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_IFE_robust_lambda.R
\name{return_robust_lambdaobject}
\alias{return_robust_lambdaobject}
\title{Calculates robust loadings}
\usage{
return_robust_lambdaobject(
  Y_like_object,
  group,
  type,
  FACTOR_GROUP = factor_group,
  number_of_group_factors = aantalfactoren_groups,
  FACTOR = comfactor,
  number_of_common_factors = aantalfactoren_common,
  NN = aantal_N,
  eclipz = FALSE
)
}
\arguments{
\item{Y_like_object}{this is Y_ster or W or W_j}

\item{group}{index of group}

\item{type}{scalar which shows in which setting this function is used}

\item{FACTOR_GROUP}{estimation of groupfactors}

\item{number_of_group_factors}{number of group factors}

\item{FACTOR}{estimation of common factors}

\item{number_of_common_factors}{number of common factors}

\item{NN}{N}

\item{eclipz}{indicator of using the eclipzdataset}
}
\value{
Nxk dataframe
}
\description{
Uses the almost classical lambda (this is an object of which the mean equals to the classical lambda) to create a robust lambda by using M estimation
}
