% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_IFE_robust_lambda.R
\name{return_robust_lambdaobject}
\alias{return_robust_lambdaobject}
\title{Calculates robust loadings}
\usage{
return_robust_lambdaobject(
  Y_like_object,
  group,
  type,
  FACTOR_GROUP = factor_group,
  AANTALFACTOREN_GROUPS = aantalfactoren_groups,
  FACTOR = comfactor,
  AANTALFACTOREN_COMMON = aantalfactoren_common
)
}
\arguments{
\item{Y_like_object}{this is Y_ster or W or W_j}

\item{group}{group}

\item{type}{scalar which shows in which setting this function is used}
}
\value{
Nxk dataframe
}
\description{
Uses the almost classical lambda to create a robust lambda by using M estimation
}
