% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_parallel.R
\name{run_config}
\alias{run_config}
\title{Wrapper around the non-parallel algorithm, to estimate beta, group membership and the factorstructures.}
\usage{
run_config(config, C_candidates, Y, X, maxit = 30)
}
\arguments{
\item{config}{contains one configuration of groups and factors}

\item{C_candidates}{candidates for C (parameter in PIC)}

\item{Y}{Y: NxT dataframe with the panel data of interest}

\item{X}{X: NxTxp array containing the observable variables}

\item{maxit}{maximum limit for the number of iterations}
}
\description{
The function estimates beta, group membership and the common and group specific factorstructures for one configuration.
}
