% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{solveFG}
\alias{solveFG}
\title{Helpfunction in update_g(), to calculate solve(FG x t(FG)) x FG}
\usage{
solveFG(TT, number_of_groups, number_of_group_factors)
}
\arguments{
\item{TT}{length of time series}

\item{number_of_groups}{number of groups}

\item{number_of_group_factors}{number of group factors to be estimated}
}
\description{
Helpfunction in update_g(), to calculate solve(FG x t(FG)) x FG
}
