% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{solveFG}
\alias{solveFG}
\title{helpfunction in update_g(), to calculate solve(FG x t(FG)) x FG}
\usage{
solveFG()
}
\description{
helpfunction in update_g(), to calculate solve(FG x t(FG)) x FG
}
