% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{test_alternative_PIC}
\alias{test_alternative_PIC}
\title{function to test alternative PIC's}
\usage{
test_alternative_PIC(term2, term3, term4)
}
\description{
function to test alternative PIC's
}
