% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_IFE_algorithm_functions.R
\name{update_g}
\alias{update_g}
\title{Function that estimates group membership.}
\usage{
update_g(
  NN = aantal_N,
  number_of_groups = aantalgroepen,
  number_of_group_factors = aantalfactoren_groups
)
}
\arguments{
\item{NN}{number of individuals}

\item{number_of_groups}{number of groups estimated}

\item{number_of_group_factors}{number of groupfactors estimated}
}
\value{
list: 1st element contains group membership and second element contains the values which are used to determine group membership
}
\description{
Function that estimates group membership.
}
\examples{
#This function needs several initial parameters to be initialized in order to work on itself.
library(RobClustTimeSeries)
library(tidyverse)
library(robustbase)
lambda_group = RobClustTimeSeries::lambda_group_real_dgp3
factor_group = RobClustTimeSeries::factor_group_real_dgp3
lambda = matrix(0,nrow=1,ncol=300) #0 factors -> but needs placeholder
comfactor = matrix(0,nrow=1,ncol=30) #0 factors -> but needs placeholder
X = RobClustTimeSeries::X_dgp3
Y = RobClustTimeSeries::Y_dgp3
g = RobClustTimeSeries::g_real_dgp3
use_robust = TRUE
b = 1 #> 0 group factors
a = 0 #0 common factors
grid = expand.grid(1:300,1:30)
homogeneous_coefficients = FALSE
heterogeneous_coefficients_groups = FALSE
heterogeneous_coefficients_individuals = TRUE
eclipz = FALSE
ABDGP1 = FALSE
ABintercept = TRUE
beta = estimate_beta(NN = 300, TT = 30, number_of_groups = 3, number_of_group_factors = c(3,3,3), number_of_common_factors = 0, number_of_variables = 3,number_vars_estimated=3,num_factors_may_vary=FALSE)[[1]]
grid = grid_add_variables(grid,beta, lambda, comfactor)
g_new = update_g(number_of_groups = 3, number_of_group_factors = c(3,3,3))[[1]]
}
