% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_X_dgp3.R
\docType{data}
\name{X_dgp3}
\alias{X_dgp3}
\title{The dataset X_dgp3 contains the values of the 3 observable variables on which Y_dgp3 is based.}
\format{
array with 300 x 30 x 3 elements
}
\usage{
X_dgp3
}
\description{
The dataset X_dgp3 contains the values of the 3 observable variables on which Y_dgp3 is based.
}
\examples{
head(X_dgp3[,,1])
hist(X_dgp3[,,1])
}
\keyword{datasets}
