% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_Y_dgp3.R
\docType{data}
\name{Y_dgp3}
\alias{Y_dgp3}
\title{Y_dgp3 contains a simulated dataset for DGP 3.}
\format{
300 x 30 matrix. Each row is one time series.
}
\usage{
Y_dgp3
}
\description{
Y = XB + LgFg. It has 3 groups and each group has 3 groupfactors.
At last there were 3 observable variables generated into it.
}
\examples{
\dontrun{
 plot(Y_dgp3[,1:2],col=g_real_dgp3)
}
}
\keyword{datasets}
